feat(ffi): define_core_ffi! — single shared source for the non-physics FFI surface
The ~250 non-physics bloom_* functions were hand-copied into six
platform crates (~9k duplicated lines) and drifted constantly: Android
shipped 60 functions behind then patched the gap with silent no-op
stubs, Windows stubbed the whole scene-graph/lighting/picking/post-FX
surface, iOS/tvOS gamepad functions took an extra leading param the
manifest doesn't declare (axis/button reads were off by one register),
and bloom_create_mesh / bloom_gen_mesh_spline_ribbon read *const f32
where Perry passes f64 arrays.

This adds the same cure the physics surface already had
(define_physics_ffi!): one macro in bloom-shared generating all 247
functions, expanded per platform. Platform crates keep only genuinely
platform-specific code (window/event loop, audio backend, clipboard,
dialogs, cursor, locale).

- new bloom_shared::ffi: panic guard at every FFI entry point (log-once
  + safe default instead of unwinding into Perry code), platform-aware
  error logging (logcat on Android), feature_off_warn_once moved here
- new bloom_shared::ffi_core: define_core_ffi! with models3d /
  image-extras gate pairs (warn-once stubs when off — symbols never
  silently vanish or no-op)
- audio::decode_audio: unified extension-dispatch + format-sniff decode
  (was two divergent per-platform behaviors)
- Renderer::set_env_clear_from_hdr_file: HDR decode hoisted from
  per-platform wrappers
- macro expansion compile-checked in shared unit tests via mock hooks
- macos migrated as reference platform: lib.rs 3567→1194 lines, release
  staticlib exports verified identical pre/post (382 symbols, 0 diff)
