+++ native/macos/src/lib.rs
@@ -9,7 +9,6 @@
 use bloom_shared::engine::EngineState;
 use bloom_shared::renderer::Renderer;
 use bloom_shared::string_header::{str_from_header, alloc_perry_string};
-use bloom_shared::audio::{parse_wav, parse_ogg, parse_mp3};
 
 use objc2::rc::Retained;
 use objc2::{msg_send, MainThreadMarker, MainThreadOnly};
@@ -31,6 +30,16 @@ static mut AUDIO_UNIT: Option<AudioUnitInstance> = None;
 fn engine() -> &'static mut EngineState {
     unsafe { ENGINE.get_mut().expect("Engine not initialized") }
 }
+/// Asset-path hook for define_core_ffi! — identity on desktop, where game
+/// asset paths are valid relative to the working directory.
+fn bloom_resolve_asset_path(path: &str) -> std::borrow::Cow<'_, str> {
+    std::borrow::Cow::Borrowed(path)
+}
+
+// The full shared (non-physics) FFI surface. See bloom_shared::ffi_core
+// docs for the contract; tools/validate-ffi.js checks parity in CI.
+bloom_shared::define_core_ffi!();
+
 
 /// Map macOS virtual key code to Bloom key code.
 fn map_keycode(keycode: u16) -> usize {
@@ -617,1406 +626,73 @@ pub extern "C" fn bloom_end_drawing() {
     engine().end_frame();
 }
 
-/// Request a PNG screenshot of the next rendered frame.
-/// The capture happens during the next end_drawing(), so the caller
-/// should call beginDrawing/endDrawing once after this for the file
-/// to actually appear on disk. Used by bloom-diff and CI image
-/// regression workflows.
-#[no_mangle]
-pub extern "C" fn bloom_take_screenshot(path_ptr: *const u8) {
-    let path = str_from_header(path_ptr).to_string();
-    let eng = engine();
-    eng.renderer.screenshot_requested = true;
-    eng.renderer.pending_screenshot_path = Some(path);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_clear_background(r: f64, g: f64, b: f64, a: f64) {
-    engine().renderer.set_clear_color(r, g, b, a);
-}
-
-#[cfg(feature = "image-extras")]
-/// Load an HDR equirectangular environment map and upload it to the
-/// GPU. Subsequent frames sample it per-background-pixel via a sky
-/// pass, so the background matches a path-traced reference instead of
-/// being a flat clear color. The file must be Radiance HDR (.hdr).
-#[no_mangle]
-pub extern "C" fn bloom_set_env_clear_from_hdr(path_ptr: *const u8) {
-    use image::ImageDecoder;
-    let path = str_from_header(path_ptr).to_string();
-    let file = match std::fs::File::open(&path) {
-        Ok(f) => f,
-        Err(_) => return,
-    };
-    let decoder = match image::codecs::hdr::HdrDecoder::new(std::io::BufReader::new(file)) {
-        Ok(d) => d,
-        Err(_) => return,
-    };
-    let (w, h) = decoder.dimensions();
-    let byte_len = (w as usize) * (h as usize) * 3 * 4;
-    let mut buf = vec![0u8; byte_len];
-    if decoder.read_image(&mut buf).is_err() {
-        return;
-    }
-    // Reinterpret the byte buffer as f32 RGB triples for the renderer.
-    let rgb_f32: Vec<f32> = buf
-        .chunks_exact(4)
-        .map(|c| f32::from_le_bytes([c[0], c[1], c[2], c[3]]))
-        .collect();
-    engine().renderer.load_env_from_hdr(w, h, &rgb_f32);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_target_fps(fps: f64) {
-    engine().target_fps = fps;
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_direct_2d_mode(on: f64) {
-    engine().direct_2d_mode = on > 0.5;
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_get_delta_time() -> f64 {
-    engine().delta_time
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_get_fps() -> f64 {
-    engine().get_fps()
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_get_screen_width() -> f64 {
-    engine().screen_width()
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_get_screen_height() -> f64 {
-    engine().screen_height()
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_get_time() -> f64 {
-    engine().get_time()
-}
-
 // ============================================================
 // Input - Keyboard
 // ============================================================
-
-#[no_mangle]
-pub extern "C" fn bloom_is_key_pressed(key: f64) -> f64 {
-    if engine().input.is_key_pressed(key as usize) { 1.0 } else { 0.0 }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_is_key_down(key: f64) -> f64 {
-    if engine().input.is_key_down(key as usize) { 1.0 } else { 0.0 }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_is_key_released(key: f64) -> f64 {
-    if engine().input.is_key_released(key as usize) { 1.0 } else { 0.0 }
-}
 
 // ============================================================
 // Input - Mouse
 // ============================================================
 
-#[no_mangle]
-pub extern "C" fn bloom_get_mouse_x() -> f64 {
-    engine().input.mouse_x
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_get_mouse_y() -> f64 {
-    engine().input.mouse_y
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_is_mouse_button_pressed(btn: f64) -> f64 {
-    if engine().input.is_mouse_button_pressed(btn as usize) { 1.0 } else { 0.0 }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_is_mouse_button_down(btn: f64) -> f64 {
-    if engine().input.is_mouse_button_down(btn as usize) { 1.0 } else { 0.0 }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_is_mouse_button_released(btn: f64) -> f64 {
-    if engine().input.is_mouse_button_released(btn as usize) { 1.0 } else { 0.0 }
-}
-
 // ============================================================
 // Input - Gamepad
 // ============================================================
-
-#[no_mangle]
-pub extern "C" fn bloom_is_gamepad_available() -> f64 {
-    if engine().input.is_gamepad_available() { 1.0 } else { 0.0 }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_get_gamepad_axis(axis: f64) -> f64 {
-    engine().input.get_gamepad_axis(axis as usize) as f64
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_is_gamepad_button_pressed(btn: f64) -> f64 {
-    if engine().input.is_gamepad_button_pressed(btn as usize) { 1.0 } else { 0.0 }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_is_gamepad_button_down(btn: f64) -> f64 {
-    if engine().input.is_gamepad_button_down(btn as usize) { 1.0 } else { 0.0 }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_is_gamepad_button_released(btn: f64) -> f64 {
-    if engine().input.is_gamepad_button_released(btn as usize) { 1.0 } else { 0.0 }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_get_gamepad_axis_count() -> f64 {
-    engine().input.get_gamepad_axis_count() as f64
-}
 
 // ============================================================
 // Input - Touch
 // ============================================================
 
-#[no_mangle]
-pub extern "C" fn bloom_get_touch_x(index: f64) -> f64 {
-    engine().input.get_touch_x(index as usize)
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_get_touch_y(index: f64) -> f64 {
-    engine().input.get_touch_y(index as usize)
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_get_touch_count() -> f64 {
-    engine().input.get_touch_count() as f64
-}
-
 // ============================================================
 // Shapes
 // ============================================================
-
-#[no_mangle]
-pub extern "C" fn bloom_draw_line(x1: f64, y1: f64, x2: f64, y2: f64, thickness: f64, r: f64, g: f64, b: f64, a: f64) {
-    engine().renderer.draw_line(x1, y1, x2, y2, thickness, r, g, b, a);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_draw_rect(x: f64, y: f64, w: f64, h: f64, r: f64, g: f64, b: f64, a: f64) {
-    engine().renderer.draw_rect(x, y, w, h, r, g, b, a);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_draw_rect_lines(x: f64, y: f64, w: f64, h: f64, thickness: f64, r: f64, g: f64, b: f64, a: f64) {
-    engine().renderer.draw_rect_lines(x, y, w, h, thickness, r, g, b, a);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_draw_circle(cx: f64, cy: f64, radius: f64, r: f64, g: f64, b: f64, a: f64) {
-    engine().renderer.draw_circle(cx, cy, radius, r, g, b, a);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_draw_circle_lines(cx: f64, cy: f64, radius: f64, r: f64, g: f64, b: f64, a: f64) {
-    engine().renderer.draw_circle_lines(cx, cy, radius, r, g, b, a);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_draw_triangle(x1: f64, y1: f64, x2: f64, y2: f64, x3: f64, y3: f64, r: f64, g: f64, b: f64, a: f64) {
-    engine().renderer.draw_triangle(x1, y1, x2, y2, x3, y3, r, g, b, a);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_draw_poly(cx: f64, cy: f64, sides: f64, radius: f64, rotation: f64, r: f64, g: f64, b: f64, a: f64) {
-    engine().renderer.draw_poly(cx, cy, sides, radius, rotation, r, g, b, a);
-}
 
 // ============================================================
 // Text
 // ============================================================
 
-#[no_mangle]
-pub extern "C" fn bloom_draw_text(text_ptr: *const u8, x: f64, y: f64, size: f64, r: f64, g: f64, b: f64, a: f64) {
-    let text = str_from_header(text_ptr);
-    let eng = engine();
-    // Need to split borrow: take text out temporarily
-    let mut text_renderer = std::mem::replace(&mut eng.text, bloom_shared::text_renderer::TextRenderer::empty());
-    text_renderer.draw_text(&mut eng.renderer, text, x, y, size as u32, r, g, b, a);
-    eng.text = text_renderer;
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_measure_text(text_ptr: *const u8, size: f64) -> f64 {
-    let text = str_from_header(text_ptr);
-    engine().text.measure_text(text, size as u32)
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_load_font(path_ptr: *const u8, _size: f64) -> f64 {
-    let path = str_from_header(path_ptr);
-    match std::fs::read(path) {
-        Ok(data) => engine().text.load_font(&data) as f64,
-        Err(_) => 0.0,
-    }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_unload_font(font_handle: f64) {
-    engine().text.unload_font(font_handle as usize);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_draw_text_ex(font_handle: f64, text_ptr: *const u8, x: f64, y: f64, size: f64, spacing: f64, r: f64, g: f64, b: f64, a: f64) {
-    let text = str_from_header(text_ptr);
-    let eng = engine();
-    let mut text_renderer = std::mem::replace(&mut eng.text, bloom_shared::text_renderer::TextRenderer::empty());
-    text_renderer.draw_text_ex(&mut eng.renderer, font_handle as usize, text, x, y, size as u32, spacing as f32, r, g, b, a);
-    eng.text = text_renderer;
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_measure_text_ex(font_handle: f64, text_ptr: *const u8, size: f64, spacing: f64) -> f64 {
-    let text = str_from_header(text_ptr);
-    engine().text.measure_text_ex(font_handle as usize, text, size as u32, spacing as f32)
-}
-
 // ============================================================
 // Textures
 // ============================================================
-
-#[no_mangle]
-pub extern "C" fn bloom_load_texture(path_ptr: *const u8) -> f64 {
-    let path = str_from_header(path_ptr);
-    match std::fs::read(path) {
-        Ok(data) => {
-            let eng = engine();
-            let EngineState { ref mut textures, ref mut renderer, .. } = *eng;
-            textures.load_texture(renderer, &data)
-        }
-        Err(_) => 0.0,
-    }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_unload_texture(handle: f64) {
-    let eng = engine();
-    let EngineState { ref mut textures, ref mut renderer, .. } = *eng;
-    textures.unload_texture(handle, renderer);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_draw_texture(handle: f64, x: f64, y: f64, tint_r: f64, tint_g: f64, tint_b: f64, tint_a: f64) {
-    let eng = engine();
-    if let Some(tex) = eng.textures.get(handle) {
-        let bind_group_idx = tex.bind_group_idx;
-        eng.renderer.draw_texture(bind_group_idx, x, y, tint_r, tint_g, tint_b, tint_a);
-    }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_draw_texture_pro(
-    handle: f64,
-    src_x: f64, src_y: f64, src_w: f64, src_h: f64,
-    dst_x: f64, dst_y: f64, dst_w: f64, dst_h: f64,
-    origin_x: f64, origin_y: f64, rotation: f64,
-    tint_r: f64, tint_g: f64, tint_b: f64, tint_a: f64,
-) {
-    let eng = engine();
-    if let Some(tex) = eng.textures.get(handle) {
-        let bind_group_idx = tex.bind_group_idx;
-        eng.renderer.draw_texture_pro(
-            bind_group_idx,
-            src_x, src_y, src_w, src_h,
-            dst_x, dst_y, dst_w, dst_h,
-            origin_x, origin_y, rotation,
-            tint_r, tint_g, tint_b, tint_a,
-        );
-    }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_draw_texture_rec(
-    handle: f64,
-    src_x: f64, src_y: f64, src_w: f64, src_h: f64,
-    dst_x: f64, dst_y: f64,
-    tint_r: f64, tint_g: f64, tint_b: f64, tint_a: f64,
-) {
-    let eng = engine();
-    if let Some(tex) = eng.textures.get(handle) {
-        let bind_group_idx = tex.bind_group_idx;
-        eng.renderer.draw_texture_rec(
-            bind_group_idx,
-            src_x, src_y, src_w, src_h,
-            dst_x, dst_y,
-            tint_r, tint_g, tint_b, tint_a,
-        );
-    }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_get_texture_width(handle: f64) -> f64 {
-    let eng = engine();
-    eng.textures.get(handle).map(|t| t.width as f64).unwrap_or(0.0)
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_get_texture_height(handle: f64) -> f64 {
-    let eng = engine();
-    eng.textures.get(handle).map(|t| t.height as f64).unwrap_or(0.0)
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_load_image(path_ptr: *const u8) -> f64 {
-    let path = str_from_header(path_ptr);
-    match std::fs::read(path) {
-        Ok(data) => engine().textures.load_image(&data),
-        Err(_) => 0.0,
-    }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_image_resize(handle: f64, w: f64, h: f64) {
-    engine().textures.image_resize(handle, w as u32, h as u32);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_image_crop(handle: f64, x: f64, y: f64, w: f64, h: f64) {
-    engine().textures.image_crop(handle, x as u32, y as u32, w as u32, h as u32);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_image_flip_h(handle: f64) {
-    engine().textures.image_flip_h(handle);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_image_flip_v(handle: f64) {
-    engine().textures.image_flip_v(handle);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_load_texture_from_image(handle: f64) -> f64 {
-    let eng = engine();
-    let EngineState { ref mut textures, ref mut renderer, .. } = *eng;
-    textures.load_texture_from_image(handle, renderer)
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_gen_texture_mipmaps(_handle: f64) {
-    // Mipmap generation is handled by the GPU texture creation pipeline
-    // This is a no-op for now as wgpu handles mipmaps internally
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_texture_filter(handle: f64, mode: f64) {
-    let eng = engine();
-    if let Some(tex) = eng.textures.get(handle) {
-        let bind_group_idx = tex.bind_group_idx;
-        eng.renderer.set_texture_filter(bind_group_idx, mode > 0.5);
-    }
-}
 
 // ============================================================
 // Camera 2D
 // ============================================================
 
-#[no_mangle]
-pub extern "C" fn bloom_begin_mode_2d(offset_x: f64, offset_y: f64, target_x: f64, target_y: f64, rotation: f64, zoom: f64) {
-    engine().renderer.begin_mode_2d(
-        offset_x as f32, offset_y as f32,
-        target_x as f32, target_y as f32,
-        rotation as f32, zoom as f32,
-    );
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_end_mode_2d() {
-    engine().renderer.end_mode_2d();
-}
-
 // ============================================================
 // Camera 3D and 3D drawing
 // ============================================================
-
-#[no_mangle]
-pub extern "C" fn bloom_begin_mode_3d(
-    pos_x: f64, pos_y: f64, pos_z: f64,
-    target_x: f64, target_y: f64, target_z: f64,
-    up_x: f64, up_y: f64, up_z: f64,
-    fovy: f64, projection: f64,
-) {
-    engine().renderer.begin_mode_3d(
-        pos_x as f32, pos_y as f32, pos_z as f32,
-        target_x as f32, target_y as f32, target_z as f32,
-        up_x as f32, up_y as f32, up_z as f32,
-        fovy as f32, projection as f32,
-    );
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_end_mode_3d() {
-    engine().renderer.end_mode_3d();
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_draw_cube(x: f64, y: f64, z: f64, w: f64, h: f64, d: f64, r: f64, g: f64, b: f64, a: f64) {
-    engine().renderer.draw_cube(x, y, z, w, h, d, r, g, b, a);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_draw_cube_wires(x: f64, y: f64, z: f64, w: f64, h: f64, d: f64, r: f64, g: f64, b: f64, a: f64) {
-    engine().renderer.draw_cube_wires(x, y, z, w, h, d, r, g, b, a);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_draw_sphere(cx: f64, cy: f64, cz: f64, radius: f64, r: f64, g: f64, b: f64, a: f64) {
-    engine().renderer.draw_sphere(cx, cy, cz, radius, r, g, b, a);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_draw_sphere_wires(cx: f64, cy: f64, cz: f64, radius: f64, r: f64, g: f64, b: f64, a: f64) {
-    engine().renderer.draw_sphere_wires(cx, cy, cz, radius, r, g, b, a);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_draw_cylinder(x: f64, y: f64, z: f64, radius_top: f64, radius_bottom: f64, height: f64, r: f64, g: f64, b: f64, a: f64) {
-    engine().renderer.draw_cylinder(x, y, z, radius_top, radius_bottom, height, r, g, b, a);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_draw_plane(cx: f64, cy: f64, cz: f64, w: f64, d: f64, r: f64, g: f64, b: f64, a: f64) {
-    engine().renderer.draw_plane(cx, cy, cz, w, d, r, g, b, a);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_draw_grid(slices: f64, spacing: f64) {
-    engine().renderer.draw_grid(slices as i32, spacing);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_draw_ray(origin_x: f64, origin_y: f64, origin_z: f64, dir_x: f64, dir_y: f64, dir_z: f64, r: f64, g: f64, b: f64, a: f64) {
-    engine().renderer.draw_ray(origin_x, origin_y, origin_z, dir_x, dir_y, dir_z, r, g, b, a);
-}
 
 // ============================================================
 // Joint test
 // ============================================================
 
-#[no_mangle]
-pub extern "C" fn bloom_set_joint_test(joint_index: f64, angle: f64) {
-    engine().renderer.set_joint_test(joint_index as usize, angle as f32);
-}
-
 // ============================================================
 // Lighting
 // ============================================================
-
-#[no_mangle]
-pub extern "C" fn bloom_set_ambient_light(r: f64, g: f64, b: f64, intensity: f64) {
-    engine().renderer.set_ambient_light(r, g, b, intensity);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_directional_light(dx: f64, dy: f64, dz: f64, r: f64, g: f64, b: f64, intensity: f64) {
-    engine().renderer.set_directional_light(dx, dy, dz, r, g, b, intensity);
-}
 
 // --- EN-005: procedural sky ---
 // Toggle procedural-atmosphere rendering and steer the sun. The
 // renderer owns the on/off flag + LUT state; setting the sun marks
 // the sky-view LUT dirty so it re-bakes before the next frame.
 
-#[no_mangle]
-pub extern "C" fn bloom_set_procedural_sky(enabled: f64, rayleigh_density: f64, mie_density: f64, ground_albedo: f64) {
-    engine().renderer.set_procedural_sky(
-        enabled != 0.0,
-        rayleigh_density as f32,
-        mie_density as f32,
-        ground_albedo as f32,
-    );
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_sun_direction(dx: f64, dy: f64, dz: f64, intensity: f64) {
-    engine().renderer.set_sun_direction(dx as f32, dy as f32, dz as f32, intensity as f32);
-}
-
 // --- Post-FX knobs (heuristic visual layer; default-off) ---
-
-#[no_mangle]
-pub extern "C" fn bloom_set_fog(r: f64, g: f64, b: f64, density: f64, height_ref: f64, height_falloff: f64) {
-    let r_ = engine();
-    r_.renderer.set_fog_color(r as f32, g as f32, b as f32);
-    r_.renderer.set_fog_density(density as f32);
-    r_.renderer.set_fog_height_falloff(height_ref as f32, height_falloff as f32);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_chromatic_aberration(strength: f64) {
-    engine().renderer.set_chromatic_aberration(strength as f32);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_vignette(strength: f64, softness: f64) {
-    engine().renderer.set_vignette(strength as f32, softness as f32);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_film_grain(strength: f64) {
-    engine().renderer.set_film_grain(strength as f32);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_sun_shafts(strength: f64, decay: f64, r: f64, g: f64, b: f64) {
-    let eng = engine();
-    eng.renderer.set_sun_shaft_strength(strength as f32);
-    eng.renderer.set_sun_shaft_decay(decay as f32);
-    eng.renderer.set_sun_shaft_color(r as f32, g as f32, b as f32);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_auto_exposure(on: f64) {
-    engine().renderer.set_auto_exposure(on != 0.0);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_taa_enabled(on: f64) {
-    engine().renderer.set_taa_enabled(on != 0.0);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_render_scale(scale: f64) {
-    engine().renderer.set_render_scale(scale as f32);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_get_render_scale() -> f64 {
-    engine().renderer.render_scale() as f64
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_upscale_mode(mode: f64) {
-    engine().renderer.set_upscale_mode(mode as u32);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_cas_strength(strength: f64) {
-    engine().renderer.set_cas_strength(strength as f32);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_get_physical_width() -> f64 {
-    engine().renderer.physical_width() as f64
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_get_physical_height() -> f64 {
-    engine().renderer.physical_height() as f64
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_auto_resolution(target_hz: f64, enabled: f64) {
-    let eng = engine();
-    if enabled != 0.0 {
-        let current = eng.renderer.render_scale();
-        eng.drs.enable(target_hz as f32, current);
-    } else {
-        eng.drs.disable();
-    }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_manual_exposure(value: f64) {
-    engine().renderer.set_manual_exposure(value as f32);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_env_intensity(intensity: f64) {
-    engine().renderer.set_env_intensity(intensity as f32);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_ssgi_enabled(enabled: f64) {
-    engine().renderer.set_ssgi_enabled(enabled != 0.0);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_ssgi_intensity(intensity: f64) {
-    engine().renderer.set_ssgi_intensity(intensity as f32);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_ssgi_radius(radius: f64) {
-    engine().renderer.set_ssgi_radius(radius as f32);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_dof(enabled: f64, focus_distance: f64, aperture: f64) {
-    let r = &mut engine().renderer;
-    r.set_dof_enabled(enabled != 0.0);
-    r.set_dof_focus_distance(focus_distance as f32);
-    r.set_dof_aperture(aperture as f32);
-}
 
 // ============================================================
 // Render quality toggles (individual + preset)
 // ============================================================
-
-#[no_mangle]
-pub extern "C" fn bloom_set_quality_preset(preset: f64) {
-    engine().renderer.apply_quality_preset(preset as u32);
-}
-#[no_mangle]
-pub extern "C" fn bloom_set_shadows_enabled(on: f64) {
-    engine().renderer.set_shadows_enabled(on != 0.0);
-}
-#[no_mangle]
-pub extern "C" fn bloom_set_shadows_always_fresh(on: f64) {
-    engine().renderer.set_shadows_always_fresh(on != 0.0);
-}
-#[no_mangle]
-pub extern "C" fn bloom_set_bloom_enabled(on: f64) {
-    engine().renderer.set_bloom_enabled(on != 0.0);
-}
-#[no_mangle]
-pub extern "C" fn bloom_set_ssao_enabled(on: f64) {
-    engine().renderer.set_ssao_enabled(on != 0.0);
-}
-#[no_mangle]
-pub extern "C" fn bloom_set_ssao_intensity(value: f64) {
-    engine().renderer.set_ssao_strength(value as f32);
-}
-#[no_mangle]
-pub extern "C" fn bloom_set_ssao_radius(world_radius: f64) {
-    engine().renderer.set_ssao_radius(world_radius as f32);
-}
-#[no_mangle]
-pub extern "C" fn bloom_set_wind(dir_x: f64, dir_z: f64, amplitude: f64, frequency: f64) {
-    engine().renderer.set_wind(dir_x as f32, dir_z as f32, amplitude as f32, frequency as f32);
-}
-#[no_mangle]
-pub extern "C" fn bloom_set_ssr_enabled(on: f64) {
-    engine().renderer.set_ssr_enabled(on != 0.0);
-}
-#[no_mangle]
-pub extern "C" fn bloom_set_motion_blur_enabled(on: f64) {
-    engine().renderer.set_motion_blur_enabled(on != 0.0);
-}
-#[no_mangle]
-pub extern "C" fn bloom_set_sss_enabled(on: f64) {
-    engine().renderer.set_sss_enabled(on != 0.0);
-}
 
 // ============================================================
 // Profiler — CPU phase timings (always available) + GPU timestamps
 // (when the adapter supports TIMESTAMP_QUERY). Disabled by default.
 // ============================================================
 
-#[no_mangle]
-pub extern "C" fn bloom_set_profiler_enabled(on: f64) {
-    engine().profiler.set_enabled(on != 0.0);
-}
-#[no_mangle]
-pub extern "C" fn bloom_get_profiler_frame_cpu_us() -> f64 {
-    engine().profiler.avg_frame_cpu_us()
-}
-#[no_mangle]
-pub extern "C" fn bloom_get_profiler_frame_gpu_us() -> f64 {
-    engine().profiler.avg_frame_gpu_us()
-}
-#[no_mangle]
-pub extern "C" fn bloom_print_profiler_summary() {
-    print!("{}", engine().profiler.summary());
-}
-
-/// Phase 8 — formatted per-pass overlay text. Returns a Perry-style
-/// header-prefixed string (12-byte header: len, cap, flags, then
-/// UTF-8 bytes) with one line per pass sorted by CPU time descending:
-///
-///   "<label>|<cpu_us>|<gpu_us_or_-1>\n..."
-///
-/// The TS overlay splits on \n then | per line. Games call this once
-/// per overlay-draw frame.
-/// Phase 7 — submit a world-space impulse. The renderer's per-frame
-/// compute pass decays + accumulates submissions into a 256×256
-/// R32Float texture covering a 128 m centred square; materials that
-/// sample `impulse_tex` (group 4 binding 4) read the result. Up to 16
-/// splats per frame are accepted — overflow is dropped silently.
-#[no_mangle]
-pub extern "C" fn bloom_splat_impulse(x: f64, z: f64, radius: f64, strength: f64) {
-    engine().renderer.impulse_field.submit_splat(
-        x as f32, z as f32, radius as f32, strength as f32,
-    );
-}
-
-/// Phase 5 — set per-material `user_params` (ABI §1.4). Bytes are
-/// uploaded verbatim to `@group(2) @binding(11)`; shaders cast them
-/// to whatever struct they declared. `param_count` is the number of
-/// f64 slots in the params array (each slot is 8 bytes in Perry's
-/// number[] layout, but we only forward the low 4 bytes as f32 since
-/// WGSL UBO scalars are 4-byte). Up to 64 floats (256 bytes).
-///
-/// Pass `param_count == 0` to revert the material to the default
-/// zero-initialised UBO.
-#[no_mangle]
-pub extern "C" fn bloom_set_material_params(
-    handle: f64,
-    params_ptr: *const f64,
-    param_count: f64,
-) {
-    let count = param_count as usize;
-    if count > 64 {
-        eprintln!("[material] set_material_params: param_count {} > 64 (256-byte UBO cap)", count);
-        return;
-    }
-    let mut bytes = vec![0u8; count * 4];
-    if !params_ptr.is_null() && count > 0 {
-        let slots = unsafe { std::slice::from_raw_parts(params_ptr, count) };
-        for (i, &v) in slots.iter().enumerate() {
-            let f = v as f32;
-            bytes[i*4..i*4+4].copy_from_slice(&f.to_le_bytes());
-        }
-    }
-    let eng = engine();
-    if let Err(e) = eng.renderer.material_system.set_user_params(
-        &eng.renderer.device, &eng.renderer.queue,
-        handle as u32, &bytes,
-    ) {
-        eprintln!("[material] set_material_params failed: {}", e);
-    }
-}
-
-/// Phase 8 — frame history for the histogram. Header-prefixed string,
-/// one line per frame "<cpu_us>|<gpu_us>", oldest first, up to 120
-/// lines.
-#[no_mangle]
-pub extern "C" fn bloom_profiler_frame_history() -> *const u8 {
-    let hist = engine().profiler.frame_history();
-    let mut s = String::with_capacity(hist.len() * 24);
-    for (cpu, gpu) in &hist {
-        s.push_str(&format!("{:.2}|{:.2}\n", cpu, gpu));
-    }
-    alloc_perry_string(&s)
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_profiler_overlay_text() -> *const u8 {
-    let snap = engine().profiler.snapshot();
-    let mut s = String::with_capacity(snap.len() * 48);
-    for (label, cpu, gpu) in &snap {
-        s.push_str(label);
-        s.push('|');
-        s.push_str(&format!("{:.2}", cpu));
-        s.push('|');
-        match gpu {
-            Some(g) => s.push_str(&format!("{:.2}", g)),
-            None    => s.push_str("-1"),
-        }
-        s.push('\n');
-    }
-    alloc_perry_string(&s)
-}
-
 // ============================================================
 // Models
 // ============================================================
 
-#[cfg(feature = "models3d")]
-#[no_mangle]
-pub extern "C" fn bloom_load_model(path_ptr: *const u8) -> f64 {
-    let path = str_from_header(path_ptr);
-    match std::fs::read(path) {
-        Ok(data) => {
-            let eng = engine();
-            let EngineState { ref mut models, ref mut renderer, .. } = *eng;
-            // .gltf = loose file with external .bin + image URIs; pass
-            // the directory so relative paths resolve. .glb (binary)
-            // embeds everything and needs no base dir.
-            let p = std::path::Path::new(&path);
-            let is_loose = p.extension().and_then(|e| e.to_str()) == Some("gltf");
-            if is_loose {
-                if let Some(dir) = p.parent() {
-                    return models.load_model_with_textures_from_path(&data, dir, renderer);
-                }
-            }
-            models.load_model_with_textures(&data, renderer)
-        }
-        Err(_) => 0.0,
-    }
-}
-
-#[cfg(feature = "models3d")]
-#[no_mangle]
-pub extern "C" fn bloom_draw_model(handle: f64, x: f64, y: f64, z: f64, scale: f64, r: f64, g: f64, b: f64, a: f64) {
-    let eng = engine();
-    if let Some(model) = eng.models.get(handle) {
-        let position = [x as f32, y as f32, z as f32];
-        let scale = scale as f32;
-        let tint = [(r / 255.0) as f32, (g / 255.0) as f32, (b / 255.0) as f32, (a / 255.0) as f32];
-        let handle_bits = handle.to_bits();
-        if eng.renderer.cache_model_if_static(handle_bits, &model.meshes) {
-            eng.renderer.draw_model_cached(handle_bits, position, scale, tint);
-        } else {
-            for mesh in &model.meshes {
-                let tex_idx = mesh.texture_idx.unwrap_or(0);
-                eng.renderer.draw_model_mesh_tinted(&mesh.vertices, &mesh.indices, position, scale, tint, tex_idx);
-            }
-        }
-    }
-}
-
-#[cfg(feature = "models3d")]
-#[no_mangle]
-pub extern "C" fn bloom_draw_model_rotated(
-    handle: f64, x: f64, y: f64, z: f64,
-    scale: f64, rot_y: f64,
-    color_packed_argb: f64,
-) {
-    // Pack RGBA into one f64 to keep the FFI to 7 args; the 9th f64
-    // would otherwise be corrupted by the Perry-ARM64 stack-arg quirk.
-    let bits = color_packed_argb as u32;
-    let a = ((bits >> 24) & 0xff) as f32 / 255.0;
-    let r = ((bits >> 16) & 0xff) as f32 / 255.0;
-    let g = ((bits >>  8) & 0xff) as f32 / 255.0;
-    let b = ( bits        & 0xff) as f32 / 255.0;
-    let eng = engine();
-    if let Some(model) = eng.models.get(handle) {
-        let position = [x as f32, y as f32, z as f32];
-        let scale = scale as f32;
-        let tint = [r, g, b, a];
-        // Skip the static-mesh cache: it bakes vertex transforms with
-        // identity rotation, so a rotated draw must take the dynamic
-        // path. Cheap — meshes are unlikely to land in the cache twice
-        // anyway when callers mix rotated + unrotated draws.
-        for mesh in &model.meshes {
-            let tex_idx = mesh.texture_idx.unwrap_or(0);
-            eng.renderer.draw_model_mesh_tinted_rotated(
-                &mesh.vertices, &mesh.indices, position, scale, tint, tex_idx, rot_y as f32,
-            );
-        }
-    }
-}
-
-#[cfg(feature = "models3d")]
-#[no_mangle]
-pub extern "C" fn bloom_unload_model(handle: f64) {
-    engine().models.unload_model(handle);
-}
-
-#[cfg(feature = "models3d")]
-#[no_mangle]
-pub extern "C" fn bloom_gen_mesh_cube(w: f64, h: f64, d: f64) -> f64 {
-    engine().models.gen_mesh_cube(w as f32, h as f32, d as f32)
-}
-
-#[cfg(feature = "models3d")]
-#[no_mangle]
-pub extern "C" fn bloom_gen_mesh_heightmap(image_handle: f64, size_x: f64, size_y: f64, size_z: f64) -> f64 {
-    let eng = engine();
-    if let Some(img) = eng.textures.images.get(image_handle) {
-        let data = img.data.clone();
-        let w = img.width;
-        let h = img.height;
-        eng.models.gen_mesh_heightmap(&data, w, h, size_x as f32, size_y as f32, size_z as f32)
-    } else {
-        0.0
-    }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_load_shader(source_ptr: *const u8) -> f64 {
-    let source = str_from_header(source_ptr);
-    engine().renderer.load_custom_shader(source) as f64
-}
-
 // ============================================================
 // Phase 1c — material system FFI
 // ============================================================
-
-#[no_mangle]
-pub extern "C" fn bloom_compile_material(source_ptr: *const u8) -> f64 {
-    let source = str_from_header(source_ptr);
-    match engine().renderer.compile_material(source) {
-        Ok(handle) => handle as f64,
-        Err(e) => {
-            eprintln!("[material] compile failed: {:?}", e);
-            0.0
-        }
-    }
-}
-
-/// Phase 4b — compile a refractive material. Shorthand for
-/// `compile_material_with_options(src, Translucent, Refractive,
-/// true)`. Perry's ARM64 FFI garbles mixed-signature calls with
-/// more than a pointer + 0-1 floats, so we ship dedicated entry
-/// points per (profile, bucket, reads_scene) combo that games
-/// actually need.
-#[no_mangle]
-pub extern "C" fn bloom_compile_material_refractive(source_ptr: *const u8) -> f64 {
-    use bloom_shared::renderer::material_pipeline::{FragmentProfile, Bucket};
-    let source = str_from_header(source_ptr);
-    match engine().renderer.compile_material_with_options(
-        source, FragmentProfile::Translucent, Bucket::Refractive, true, false,
-    ) {
-        Ok(handle) => handle as f64,
-        Err(e) => { eprintln!("[refractive] compile failed: {:?}", e); 0.0 }
-    }
-}
-
-/// Phase 4b — compile a transparent (non-refractive) material.
-/// Profile=Translucent, Bucket=Transparent, reads_scene=false.
-#[no_mangle]
-pub extern "C" fn bloom_compile_material_transparent(source_ptr: *const u8) -> f64 {
-    use bloom_shared::renderer::material_pipeline::{FragmentProfile, Bucket};
-    let source = str_from_header(source_ptr);
-    match engine().renderer.compile_material_with_options(
-        source, FragmentProfile::Translucent, Bucket::Transparent, false, false,
-    ) {
-        Ok(handle) => handle as f64,
-        Err(e) => { eprintln!("[material] compile failed: {:?}", e); 0.0 }
-    }
-}
-
-/// Phase 4b — compile an additive material. Profile=Translucent,
-/// Bucket=Additive, reads_scene=false.
-#[no_mangle]
-pub extern "C" fn bloom_compile_material_additive(source_ptr: *const u8) -> f64 {
-    use bloom_shared::renderer::material_pipeline::{FragmentProfile, Bucket};
-    let source = str_from_header(source_ptr);
-    match engine().renderer.compile_material_with_options(
-        source, FragmentProfile::Translucent, Bucket::Additive, false, false,
-    ) {
-        Ok(handle) => handle as f64,
-        Err(e) => { eprintln!("[material] compile failed: {:?}", e); 0.0 }
-    }
-}
-
-/// EN-010 — compile a cutout (alpha-tested) material. Profile=Opaque
-/// (full G-buffer write, sun shadow, SSAO), Bucket=Cutout (rendered
-/// double-sided so foliage is visible from both sides),
-/// reads_scene=false. The fragment shader is expected to call
-/// `discard` against `material.metal_rough.w`
-/// (`MaterialFactors.alpha_cutoff`) to drop transparent texels.
-#[no_mangle]
-pub extern "C" fn bloom_compile_material_cutout(source_ptr: *const u8) -> f64 {
-    use bloom_shared::renderer::material_pipeline::{FragmentProfile, Bucket};
-    let source = str_from_header(source_ptr);
-    match engine().renderer.compile_material_with_options(
-        source, FragmentProfile::Opaque, Bucket::Cutout, false, false,
-    ) {
-        Ok(handle) => handle as f64,
-        Err(e) => { eprintln!("[material] compile failed: {:?}", e); 0.0 }
-    }
-}
-
-/// EN-001 — compile a material pipeline that opts into the standard
-/// per-instance vertex layout (Opaque profile + Opaque bucket +
-/// wants_instancing). Pair with `bloom_create_instance_buffer` +
-/// `bloom_submit_material_draw_instanced` to drive per-instance
-/// rendering through a single draw_indexed call.
-///
-/// Like the other `bloom_compile_material_*` variants this is
-/// macOS-only: the existing material-compile FFI surface lives only
-/// on macOS today (per the EN-010 precedent), and porting all five
-/// variants to ios/tvos/windows/linux/android/web is out of scope for
-/// EN-001. The runtime triple
-/// (`bloom_create_instance_buffer` / `bloom_submit_material_draw_instanced`
-/// / `bloom_destroy_instance_buffer`) follows the same scope so the
-/// surface is consistent — they're useless without the compile FFI.
-#[no_mangle]
-pub extern "C" fn bloom_compile_material_instanced(source_ptr: *const u8) -> f64 {
-    let source = str_from_header(source_ptr);
-    match engine().renderer.compile_material_instanced(source) {
-        Ok(handle) => handle as f64,
-        Err(e) => { eprintln!("[material] instanced compile failed: {:?}", e); 0.0 }
-    }
-}
-
-/// EN-001 — upload a flat per-instance buffer to GPU memory. Perry's
-/// `number[]` is f64-laid-out in memory; this FFI takes the pointer
-/// + the instance count and reads `instance_count * 9` f64 slots.
-/// The Rust side downcasts to f32 and pads each instance to 12 floats
-/// (48 bytes, vec4-aligned) for the GPU stride. Returns a 1-based
-/// handle, or 0 on failure (null pointer or zero count).
-#[no_mangle]
-pub extern "C" fn bloom_create_instance_buffer(
-    data_ptr: *const f64, instance_count: f64,
-) -> f64 {
-    if data_ptr.is_null() || instance_count <= 0.0 { return 0.0; }
-    let count = instance_count as u32;
-    let slot_count = (count as usize) * 9;
-    let raw_f64 = unsafe { std::slice::from_raw_parts(data_ptr, slot_count) };
-    let raw_f32: Vec<f32> = raw_f64.iter().map(|&v| v as f32).collect();
-    engine().renderer.create_instance_buffer(&raw_f32, count) as f64
-}
-
-#[cfg(feature = "models3d")]
-/// EN-001 — submit an instanced material draw. The mesh at
-/// (mesh_handle, mesh_idx) is drawn `instance_count` times via a
-/// single `draw_indexed` with the instance buffer bound at vertex
-/// slot 1. The pipeline at `material` must have been compiled via
-/// `bloom_compile_material_instanced`. Capped at 5 f64 args to dodge
-/// the Perry-ARM64 9th-arg garbling quirk.
-#[no_mangle]
-pub extern "C" fn bloom_submit_material_draw_instanced(
-    material: f64, mesh_handle: f64, mesh_idx: f64,
-    instance_buffer: f64, instance_count: f64,
-) {
-    let eng = engine();
-    let handle_bits = mesh_handle.to_bits();
-    if let Some(model) = eng.models.get(mesh_handle) {
-        eng.renderer.cache_model_if_static(handle_bits, &model.meshes);
-    }
-    eng.renderer.submit_material_draw_instanced(
-        material as u32,
-        handle_bits,
-        mesh_idx as usize,
-        instance_buffer as u32,
-        instance_count as u32,
-    );
-}
-
-/// EN-001 — release the GPU memory backing an instance buffer. The
-/// slot stays `None` so previously-issued handles can never alias a
-/// future allocation. Safe to call with handle 0 or stale handles.
-#[no_mangle]
-pub extern "C" fn bloom_destroy_instance_buffer(handle: f64) {
-    engine().renderer.destroy_instance_buffer(handle as u32);
-}
-
-/// EN-011 — create a planar reflection probe and return its 1-based
-/// handle (0 on failure). Each frame, the engine renders the world
-/// (minus materials linked to a probe) from a camera mirrored across
-/// (`plane_y`, `normal_xyz`) into a square HDR RT at `resolution²`.
-/// Pass `resolution = 0` to default to half the swapchain width.
-///
-/// Pair with `bloom_set_material_reflection_probe` to wire the probe
-/// into a water material's `@group(2) @binding(12)` slot.
-#[no_mangle]
-pub extern "C" fn bloom_create_planar_reflection(
-    plane_y: f64, nx: f64, ny: f64, nz: f64, resolution: f64,
-) -> f64 {
-    engine().renderer.create_planar_reflection(
-        plane_y as f32,
-        [nx as f32, ny as f32, nz as f32],
-        resolution as u32,
-    ) as f64
-}
-
-/// EN-011 — link a material to a planar reflection probe. Subsequent
-/// draws of `material` see the probe's RT at `@group(2) @binding(12)`.
-/// Pass `probe = 0` to revert to the default 1×1 black texture.
-#[no_mangle]
-pub extern "C" fn bloom_set_material_reflection_probe(
-    material: f64, probe: f64,
-) {
-    engine().renderer.set_material_reflection_probe(material as u32, probe as u32);
-}
-
-/// EN-014 — create a texture array from concatenated RGBA8 byte data.
-/// `data_ptr` points to `layer_count` × `width` × `height` × 4 bytes:
-/// each layer is one slab back-to-back. All layers share the same
-/// `width` × `height` (wgpu requires uniform extent for D2Array).
-/// Layer count is capped at 16 in V1; extras are dropped.
-///
-/// Returns a 1-based handle (0 on failure: null pointer, zero count,
-/// zero extent, or short data buffer). Pair with
-/// `bloom_set_material_texture_array` to bind the array to a
-/// material's @group(2) @binding 14/15/16 slot.
-///
-/// V1 takes raw bytes (not pre-loaded texture handles) because
-/// existing `bloom_load_texture` uploads to GPU and drops the source
-/// bytes — there's no CPU-side image registry to read from. Games
-/// pre-decode their layer images via the standard image-loading API
-/// (`loadImage`, etc.) and pass the raw RGBA bytes here.
-#[no_mangle]
-pub extern "C" fn bloom_create_texture_array(
-    data_ptr:    *const u8,
-    data_len:    f64,
-    width:       f64,
-    height:      f64,
-    layer_count: f64,
-) -> f64 {
-    // EN-014 V2 — V1 stays callable; forwards to _ex with default
-    // format = sRGB (0) and mip_levels = 1 (no mips).
-    bloom_create_texture_array_ex(data_ptr, data_len, width, height, layer_count, 0.0, 1.0)
-}
-
-/// EN-014 V2 — create a texture array with explicit format + mip control.
-/// See `MaterialSystem::create_texture_array_ex` for semantics. The V1
-/// entry above is preserved as a thin wrapper.
-#[no_mangle]
-pub extern "C" fn bloom_create_texture_array_ex(
-    data_ptr:    *const u8,
-    data_len:    f64,
-    width:       f64,
-    height:      f64,
-    layer_count: f64,
-    format:      f64,
-    mip_levels:  f64,
-) -> f64 {
-    if data_ptr.is_null() || data_len <= 0.0 { return 0.0; }
-    let w = width as u32;
-    let h = height as u32;
-    if w == 0 || h == 0 { return 0.0; }
-    let layers_count = (layer_count as u32)
-        .min(bloom_shared::renderer::material_system::MAX_TEXTURE_ARRAY_LAYERS);
-    if layers_count == 0 { return 0.0; }
-    let layer_size = (w as usize) * (h as usize) * 4;
-    let total_bytes = (data_len as usize)
-        .min(layers_count as usize * layer_size);
-    let bytes = unsafe { std::slice::from_raw_parts(data_ptr, total_bytes) };
-    let mut layers: Vec<(&[u8], u32, u32)> = Vec::with_capacity(layers_count as usize);
-    for i in 0..(layers_count as usize) {
-        let start = i * layer_size;
-        let end   = start + layer_size;
-        if end > bytes.len() { break; }
-        layers.push((&bytes[start..end], w, h));
-    }
-    engine().renderer.create_texture_array_ex(&layers, format as u32, mip_levels as u32) as f64
-}
-
-/// EN-014 — link a texture-array handle to a material at one of three
-/// slots: 0 = albedo (binding 14), 1 = normal (binding 15),
-/// 2 = MR (binding 16). Pass `array = 0` to revert to the engine's
-/// 1×1×1 stub. No-op for unknown handles or out-of-range slots.
-#[no_mangle]
-pub extern "C" fn bloom_set_material_texture_array(
-    material: f64, slot: f64, array: f64,
-) {
-    engine().renderer.set_material_texture_array(
-        material as u32, slot as u32, array as u32,
-    );
-}
-
-/// EN-012 — set the shading model for a material.
-/// `0` = default lit (standard PBR),
-/// `1` = foliage (wrap-lambert + transmission via `shade_foliage`),
-/// `2` = subsurface (V2 stub — currently behaves as default lit).
-/// Lazily allocates a per-material `MaterialFactors` UBO on first
-/// call. Game shaders branch on `material.shading_model.x` and call
-/// either `shade_pbr_standard` or `shade_foliage` from the engine
-/// helper in `common/pbr.wgsl`.
-#[no_mangle]
-pub extern "C" fn bloom_set_material_shading_model(
-    material: f64, model: f64,
-) {
-    engine().renderer.set_material_shading_model(material as u32, model as u32);
-}
-
-/// EN-012 — set the foliage shading parameters for a material.
-/// Only takes effect when the material's `shading_model == 1`.
-///
-/// `trans_r/g/b`: rgb tint for back-lit foliage (1,1,1 = neutral).
-/// `trans_amount`: 0..1 — how much sun bleeds through the leaf when
-///                 the camera looks at it backlit by the sun.
-/// `wrap_factor`: 0..1 — wrap-lambert intensity. 0 = standard lambert
-///                 (back faces black), 1 = light wraps fully around.
-#[no_mangle]
-pub extern "C" fn bloom_set_material_foliage(
-    material: f64,
-    trans_r: f64, trans_g: f64, trans_b: f64,
-    trans_amount: f64, wrap_factor: f64,
-) {
-    engine().renderer.set_material_foliage(
-        material as u32,
-        [trans_r as f32, trans_g as f32, trans_b as f32],
-        trans_amount as f32, wrap_factor as f32,
-    );
-}
-
-/// Phase 6 — file-backed material compile. The path is registered
-/// with the hot-reload watcher; subsequent edits trigger an automatic
-/// recompile on the next end_frame. `bucket_kind` selects the
-/// (profile, bucket, reads_scene) preset:
-///   0 = opaque
-///   1 = transparent (translucent, no scene snapshot)
-///   2 = refractive  (translucent + scene snapshot)
-///   3 = additive    (translucent, no scene snapshot, additive blend)
-///   4 = cutout      (opaque profile, double-sided, fragment-discard
-///                    against MaterialFactors.alpha_cutoff)
-#[no_mangle]
-pub extern "C" fn bloom_compile_material_from_file(
-    path_ptr: *const u8,
-    bucket_kind: f64,
-) -> f64 {
-    use bloom_shared::renderer::material_pipeline::{FragmentProfile, Bucket};
-    let path = str_from_header(path_ptr);
-    let (profile, bucket, reads_scene) = match bucket_kind as u32 {
-        0 => (FragmentProfile::Opaque,      Bucket::Opaque,      false),
-        1 => (FragmentProfile::Translucent, Bucket::Transparent, false),
-        2 => (FragmentProfile::Translucent, Bucket::Refractive,  true),
-        3 => (FragmentProfile::Translucent, Bucket::Additive,    false),
-        4 => (FragmentProfile::Opaque,      Bucket::Cutout,      false),
-        _ => {
-            eprintln!("[material] from_file: unknown bucket_kind {bucket_kind}");
-            return 0.0;
-        }
-    };
-    match engine().renderer.compile_material_from_file(
-        std::path::Path::new(path), profile, bucket, reads_scene,
-    ) {
-        Ok(handle) => handle as f64,
-        Err(e) => { eprintln!("[material] from_file failed: {e}"); 0.0 }
-    }
-}
-
-/// EN-017 — compile + install a fullscreen post-pass material. Runs
-/// after composite + tonemapping, before the 2D overlay (HUD stays
-/// crisp). The fragment shader must declare:
-///
-///     @fragment fn fs_main(@location(0) uv: vec2<f32>)
-///         -> @location(0) vec4<f32> { ... }
-///
-/// and may sample `scene_color_tex` (LDR, post-tonemap), `scene_depth_tex`
-/// from the engine-provided bind group at `@group(0)` —
-/// see `post_pass::POST_PASS_PRELUDE` for the full ABI. Returns 1.0
-/// on success, 0.0 on compile failure.
-#[no_mangle]
-pub extern "C" fn bloom_set_post_pass(source_ptr: *const u8) -> f64 {
-    let source = str_from_header(source_ptr);
-    match engine().renderer.set_post_pass(source) {
-        Ok(()) => 1.0,
-        Err(e) => { eprintln!("[post_pass] compile failed: {:?}", e); 0.0 }
-    }
-}
-
-/// EN-017 — uninstall the active post-pass. Composite goes back to
-/// writing the swapchain directly (zero-cost original path).
-#[no_mangle]
-pub extern "C" fn bloom_clear_post_pass() {
-    engine().renderer.clear_post_pass();
-}
-
-/// EN-017 V2 — append a post-pass to the stack. Returns a 1-based
-/// handle on success (so 0 means failure at the FFI layer; the TS
-/// wrapper subtracts 1 to expose a 0-based index). The fragment
-/// shader sees `scene_color_tex` (LDR, post-tonemap) +
-/// `scene_depth_tex` at `@group(0)` — same ABI as
-/// `bloom_set_post_pass`, just cumulative.
-#[no_mangle]
-pub extern "C" fn bloom_add_post_pass(source_ptr: *const u8) -> f64 {
-    let source = str_from_header(source_ptr);
-    match engine().renderer.add_post_pass(source) {
-        Ok(h) => h as f64,
-        Err(e) => { eprintln!("[post_pass] compile failed: {:?}", e); 0.0 }
-    }
-}
-
-/// EN-017 V2 — wipe the entire post-pass stack. Composite goes back
-/// to writing the swapchain directly (zero post-pass cost).
-#[no_mangle]
-pub extern "C" fn bloom_clear_all_post_passes() {
-    engine().renderer.clear_all_post_passes();
-}
-
-#[cfg(feature = "models3d")]
-#[no_mangle]
-pub extern "C" fn bloom_draw_material(
-    material: f64,
-    mesh_handle: f64,
-    mesh_idx: f64,
-    x: f64, y: f64, z: f64, scale: f64,
-    r: f64, g: f64, b: f64, a: f64,
-) {
-    let eng = engine();
-    let handle_bits = mesh_handle.to_bits();
-    if let Some(model) = eng.models.get(mesh_handle) {
-        eng.renderer.cache_model_if_static(handle_bits, &model.meshes);
-    }
-    eng.renderer.submit_material_draw(
-        material as u32,
-        handle_bits,
-        mesh_idx as usize,
-        [x as f32, y as f32, z as f32],
-        scale as f32,
-        [(r / 255.0) as f32, (g / 255.0) as f32, (b / 255.0) as f32, (a / 255.0) as f32],
-    );
-}
-
-#[cfg(feature = "models3d")]
-#[no_mangle]
-pub extern "C" fn bloom_create_mesh(vertex_ptr: *const f64, vertex_count: f64, index_ptr: *const f64, index_count: f64) -> f64 {
-    // Perry's TS `number[]` is f64-laid-out in memory; Perry passes a
-    // pointer to that data. A previous version of this FFI declared
-    // *const f32 / *const u32, which silently read the low 4 bytes
-    // of each f64 slot as garbage f32/u32 values — meshes registered
-    // (non-zero handle) but were unrenderable.
-    //
-    // Caller must pass `vertex_count` and `index_count` derived from
-    // a literal-initialized array OR from values it computed itself.
-    // Don't compute these via `arr.length` after `.push()` — Perry's
-    // `.length` property currently reflects the literal-init size,
-    // not the post-push count (a Perry codegen bug). Workaround on
-    // the TS side: track the count manually or use literal arrays.
-    if vertex_ptr.is_null() || index_ptr.is_null() { return 0.0; }
-    let vcount = vertex_count as usize;
-    let icount = index_count as usize;
-    let vertex_f64 = unsafe { std::slice::from_raw_parts(vertex_ptr, vcount * 12) };
-    let index_f64 = unsafe { std::slice::from_raw_parts(index_ptr, icount) };
-    let vertex_data: Vec<f32> = vertex_f64.iter().map(|&v| v as f32).collect();
-    let index_data: Vec<u32> = index_f64.iter().map(|&v| v as u32).collect();
-    engine().models.create_mesh(&vertex_data, &index_data)
-}
-
-#[cfg(feature = "models3d")]
-#[no_mangle]
-pub extern "C" fn bloom_load_model_animation(path_ptr: *const u8) -> f64 {
-    let path = str_from_header(path_ptr);
-    match std::fs::read(path) {
-        Ok(data) => engine().models.load_model_animation(&data),
-        Err(_) => 0.0,
-    }
-}
-
-#[cfg(feature = "models3d")]
-#[no_mangle]
-pub extern "C" fn bloom_update_model_animation(handle: f64, anim_index: f64, time: f64, scale: f64, px: f64, py: f64, pz: f64, rot_y: f64) {
-    // Take a single Y-axis angle (radians) instead of sin/cos, so the
-    // engine reconstructs both with full precision + correct signs.
-    // Older callers that passed (rot_sin, rot_cos) hit a Perry-ARM64
-    // 9th-arg garbling bug AND a sqrt(1-sin²) workaround that lost
-    // the sign of cos — model rotation was correct only on half the
-    // circle. 8-arg signature dodges both issues.
-    let rot_y_f = rot_y as f32;
-    let rot_sin = rot_y_f.sin();
-    let rot_cos = rot_y_f.cos();
-    let eng = engine();
-    eng.models.update_model_animation(handle, anim_index as usize, time as f32);
-    if let Some(anim) = eng.models.get_animation(handle) {
-        if !anim.joint_matrices.is_empty() {
-            eng.renderer.set_joint_matrices_scaled(&anim.joint_matrices, scale as f32, [px as f32, py as f32, pz as f32], rot_sin, rot_cos);
-        }
-    }
-}
-
-#[cfg(feature = "models3d")]
-#[no_mangle]
-pub extern "C" fn bloom_get_model_mesh_count(handle: f64) -> f64 {
-    match engine().models.get(handle) {
-        Some(model) => model.meshes.len() as f64,
-        None => 0.0,
-    }
-}
-
-#[cfg(feature = "models3d")]
-#[no_mangle]
-pub extern "C" fn bloom_get_model_material_count(handle: f64) -> f64 {
-    match engine().models.get(handle) {
-        Some(model) => model.meshes.len() as f64, // materials roughly equal meshes
-        None => 0.0,
-    }
-}
 
 // ============================================================
 // Audio
@@ -2097,108 +773,9 @@ pub extern "C" fn bloom_close_audio() {
     }
 }
 
-#[no_mangle]
-pub extern "C" fn bloom_load_sound(path_ptr: *const u8) -> f64 {
-    let path = str_from_header(path_ptr);
-    match std::fs::read(path) {
-        Ok(data) => {
-            let sound_data = if path.ends_with(".ogg") || path.ends_with(".OGG") {
-                parse_ogg(&data)
-            } else if path.ends_with(".mp3") || path.ends_with(".MP3") {
-                parse_mp3(&data)
-            } else {
-                parse_wav(&data)
-            };
-            if let Some(sound_data) = sound_data {
-                engine().audio.load_sound(sound_data)
-            } else {
-                0.0
-            }
-        }
-        Err(_) => 0.0,
-    }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_play_sound(handle: f64) {
-    engine().audio.play_sound(handle);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_stop_sound(handle: f64) {
-    engine().audio.stop_sound(handle);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_sound_volume(handle: f64, volume: f64) {
-    engine().audio.set_sound_volume(handle, volume as f32);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_master_volume(volume: f64) {
-    engine().audio.master_volume = volume as f32;
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_play_sound_3d(handle: f64, x: f64, y: f64, z: f64) {
-    engine().audio.play_sound_3d(handle, x as f32, y as f32, z as f32);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_listener_position(x: f64, y: f64, z: f64, fx: f64, fy: f64, fz: f64) {
-    engine().audio.set_listener_position(x as f32, y as f32, z as f32, fx as f32, fy as f32, fz as f32);
-}
-
 // ============================================================
 // Music
 // ============================================================
-
-#[no_mangle]
-pub extern "C" fn bloom_load_music(path_ptr: *const u8) -> f64 {
-    let path = str_from_header(path_ptr);
-    match std::fs::read(path) {
-        Ok(data) => {
-            let sound_data = if path.ends_with(".ogg") || path.ends_with(".OGG") {
-                parse_ogg(&data)
-            } else if path.ends_with(".mp3") || path.ends_with(".MP3") {
-                parse_mp3(&data)
-            } else {
-                parse_wav(&data)
-            };
-            if let Some(sound_data) = sound_data {
-                engine().audio.load_music(sound_data)
-            } else {
-                0.0
-            }
-        }
-        Err(_) => 0.0,
-    }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_play_music(handle: f64) {
-    engine().audio.play_music(handle);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_stop_music(handle: f64) {
-    engine().audio.stop_music(handle);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_update_music_stream(handle: f64) {
-    engine().audio.update_music_stream(handle);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_set_music_volume(handle: f64, volume: f64) {
-    engine().audio.set_music_volume(handle, volume as f32);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_is_music_playing(handle: f64) -> f64 {
-    if engine().audio.is_music_playing(handle) { 1.0 } else { 0.0 }
-}
 
 // ============================================================
 // Utility
@@ -2267,38 +844,6 @@ pub extern "C" fn bloom_enable_cursor() {
     }
 }
 
-#[no_mangle]
-pub extern "C" fn bloom_get_mouse_delta_x() -> f64 {
-    engine().input.mouse_delta_x
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_get_mouse_delta_y() -> f64 {
-    engine().input.mouse_delta_y
-}
-
-// Accumulated scroll wheel delta since the last call. Reading consumes the
-// value (returns 0 on the next call until the user scrolls again). Used by
-// the editor's orbit camera and any scrollable UI panel.
-#[no_mangle]
-pub extern "C" fn bloom_get_mouse_wheel() -> f64 {
-    engine().input.consume_mouse_wheel()
-}
-
-// Dequeue the next typed character (Unicode codepoint). Returns 0 when the
-// queue is empty. Call in a loop each frame to consume all pending characters.
-// Used by the editor's in-window text-input widget.
-#[no_mangle]
-pub extern "C" fn bloom_get_char_pressed() -> f64 {
-    engine().input.pop_char() as f64
-}
-
-// Q2: Cursor shape
-#[no_mangle]
-pub extern "C" fn bloom_set_cursor_shape(shape: f64) {
-    engine().input.cursor_shape = shape as u32;
-}
-
 // E4: Clipboard
 #[no_mangle]
 pub extern "C" fn bloom_set_clipboard_text(text_ptr: *const u8) {
@@ -2347,92 +892,9 @@ pub extern "C" fn bloom_save_file_dialog(default_name_ptr: *const u8, title_ptr:
     }
 }
 
-// Model bounds accessors. Return the axis-aligned bounding box of a loaded
-// model in model-local coordinates. Editors use these to size gizmos, auto-
-// frame the camera on selection, and snap placed entities onto terrain.
-#[cfg(feature = "models3d")]
-#[no_mangle]
-pub extern "C" fn bloom_get_model_bounds_min_x(model_handle: f64) -> f64 {
-    engine().models.get_bounds(model_handle).0[0] as f64
-}
-#[cfg(feature = "models3d")]
-#[no_mangle]
-pub extern "C" fn bloom_get_model_bounds_min_y(model_handle: f64) -> f64 {
-    engine().models.get_bounds(model_handle).0[1] as f64
-}
-#[cfg(feature = "models3d")]
-#[no_mangle]
-pub extern "C" fn bloom_get_model_bounds_min_z(model_handle: f64) -> f64 {
-    engine().models.get_bounds(model_handle).0[2] as f64
-}
-#[cfg(feature = "models3d")]
-#[no_mangle]
-pub extern "C" fn bloom_get_model_bounds_max_x(model_handle: f64) -> f64 {
-    engine().models.get_bounds(model_handle).1[0] as f64
-}
-#[cfg(feature = "models3d")]
-#[no_mangle]
-pub extern "C" fn bloom_get_model_bounds_max_y(model_handle: f64) -> f64 {
-    engine().models.get_bounds(model_handle).1[1] as f64
-}
-#[cfg(feature = "models3d")]
-#[no_mangle]
-pub extern "C" fn bloom_get_model_bounds_max_z(model_handle: f64) -> f64 {
-    engine().models.get_bounds(model_handle).1[2] as f64
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_write_file(path_ptr: *const u8, data_ptr: *const u8) -> f64 {
-    let path = str_from_header(path_ptr);
-    let data = str_from_header(data_ptr);
-    match std::fs::write(path, data.as_bytes()) {
-        Ok(_) => 1.0,
-        Err(_) => 0.0,
-    }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_file_exists(path_ptr: *const u8) -> f64 {
-    let path = str_from_header(path_ptr);
-    if std::path::Path::new(path).exists() { 1.0 } else { 0.0 }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_read_file(path_ptr: *const u8) -> *const u8 {
-    let path = str_from_header(path_ptr);
-    // Always return a valid Perry string. A null pointer would NaN-box into a
-    // string-typed JS value pointing at address 0; `.length` / `.charCodeAt`
-    // would then segfault. Callers detect "missing file" via `data.length === 0`.
-    match std::fs::read_to_string(path) {
-        Ok(contents) => alloc_perry_string(&contents),
-        Err(_)       => alloc_perry_string(""),
-    }
-}
-
 // ============================================================
 // Input injection + platform detection
 // ============================================================
-
-#[no_mangle]
-pub extern "C" fn bloom_inject_key_down(key: f64) {
-    engine().input.set_key_down(key as usize);
-}
-#[no_mangle]
-pub extern "C" fn bloom_inject_key_up(key: f64) {
-    engine().input.set_key_up(key as usize);
-}
-#[no_mangle]
-pub extern "C" fn bloom_inject_gamepad_axis(axis: f64, value: f64) {
-    engine().input.set_gamepad_axis(axis as usize, value as f32);
-}
-#[no_mangle]
-pub extern "C" fn bloom_inject_gamepad_button_down(button: f64) {
-    engine().input.set_gamepad_button_down(button as usize);
-}
-#[no_mangle]
-pub extern "C" fn bloom_inject_gamepad_button_up(button: f64) {
-    engine().input.set_gamepad_button_up(button as usize);
-}
 #[no_mangle]
 pub extern "C" fn bloom_get_platform() -> f64 { 1.0 }
 
@@ -2454,428 +916,34 @@ pub extern "C" fn bloom_get_language() -> f64 {
         None => pack("en"),
     }
 }
-#[no_mangle]
-pub extern "C" fn bloom_is_any_input_pressed() -> f64 {
-    if engine().input.is_any_input_pressed() { 1.0 } else { 0.0 }
-}
-#[no_mangle]
-pub extern "C" fn bloom_get_crown_rotation() -> f64 {
-    engine().input.consume_crown_rotation()
-}
 
 // ============================================================
 // Frame callbacks
 // ============================================================
 
-#[no_mangle]
-pub extern "C" fn bloom_register_frame_callback(priority: f64, callback: extern "C" fn(f64)) -> f64 {
-    engine().frame_callbacks.register(priority as i32, callback) as f64
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_unregister_frame_callback(id: f64) {
-    engine().frame_callbacks.unregister(id as u64);
-}
-
-/// Emscripten-style game loop. On native, this is a no-op — the game uses
-/// a while(!windowShouldClose()) loop in TypeScript instead. The TS-level
-/// runGame() helper handles the native loop. Only used on web where the
-/// JS glue intercepts this and drives requestAnimationFrame.
-#[no_mangle]
-pub extern "C" fn bloom_run_game(_callback: extern "C" fn(f64)) {
-    // No-op on native. The TypeScript runGame() helper provides the while loop.
-}
-
 // ============================================================
 // Multiple lights
 // ============================================================
-
-#[no_mangle]
-pub extern "C" fn bloom_add_directional_light(
-    dx: f64, dy: f64, dz: f64,
-    r: f64, g: f64, b: f64,
-    intensity: f64,
-) {
-    engine().renderer.add_directional_light(
-        dx as f32, dy as f32, dz as f32,
-        r as f32, g as f32, b as f32,
-        intensity as f32,
-    );
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_add_point_light(
-    x: f64, y: f64, z: f64, range: f64,
-    r: f64, g: f64, b: f64,
-    intensity: f64,
-) {
-    engine().renderer.add_point_light(
-        x as f32, y as f32, z as f32, range as f32,
-        r as f32, g as f32, b as f32,
-        intensity as f32,
-    );
-}
 
 // ============================================================
 // Scene graph (retained mode)
 // ============================================================
 
-#[no_mangle]
-pub extern "C" fn bloom_scene_create_node() -> f64 {
-    engine().scene.create_node()
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_scene_destroy_node(handle: f64) {
-    engine().scene.destroy_node(handle);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_scene_set_visible(handle: f64, visible: f64) {
-    engine().scene.set_visible(handle, visible != 0.0);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_scene_set_cast_shadow(handle: f64, cast: f64) {
-    engine().scene.set_cast_shadow(handle, cast != 0.0);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_scene_set_receive_shadow(handle: f64, receive: f64) {
-    engine().scene.set_receive_shadow(handle, receive != 0.0);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_scene_set_parent(handle: f64, parent: f64) {
-    engine().scene.set_parent(handle, parent);
-}
-
-/// Set 4x4 transform matrix (16 floats passed as pointer to f64 array).
-#[no_mangle]
-pub extern "C" fn bloom_scene_set_transform(handle: f64, mat_ptr: *const f64) {
-    if mat_ptr.is_null() { return; }
-    let slice = unsafe { std::slice::from_raw_parts(mat_ptr, 16) };
-    let mut mat = [[0.0f32; 4]; 4];
-    for col in 0..4 {
-        for row in 0..4 {
-            mat[col][row] = slice[col * 4 + row] as f32;
-        }
-    }
-    engine().scene.set_transform(handle, mat);
-}
-
-/// Update geometry from flat vertex array (12 floats/vertex: xyz, nxnynz, rgba, uv)
-/// and index array.
-#[no_mangle]
-pub extern "C" fn bloom_scene_update_geometry(
-    handle: f64,
-    vert_ptr: *const f64,
-    vert_count: f64,
-    idx_ptr: *const f64,
-    idx_count: f64,
-) {
-    if vert_ptr.is_null() || idx_ptr.is_null() { return; }
-    let nv = vert_count as usize;
-    let ni = idx_count as usize;
-
-    let vert_floats = unsafe { std::slice::from_raw_parts(vert_ptr, nv * 12) };
-    let idx_floats = unsafe { std::slice::from_raw_parts(idx_ptr, ni) };
-
-    let mut vertices = Vec::with_capacity(nv);
-    for i in 0..nv {
-        let base = i * 12;
-        vertices.push(bloom_shared::renderer::Vertex3D {
-            position: [vert_floats[base] as f32, vert_floats[base+1] as f32, vert_floats[base+2] as f32],
-            normal: [vert_floats[base+3] as f32, vert_floats[base+4] as f32, vert_floats[base+5] as f32],
-            color: [vert_floats[base+6] as f32, vert_floats[base+7] as f32, vert_floats[base+8] as f32, vert_floats[base+9] as f32],
-            uv: [vert_floats[base+10] as f32, vert_floats[base+11] as f32],
-            joints: [0.0; 4],
-            weights: [0.0; 4],
-            tangent: [0.0; 4],
-        });
-    }
-
-    let indices: Vec<u32> = idx_floats.iter().map(|&v| v as u32).collect();
-
-    engine().scene.update_geometry(handle, vertices, indices);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_scene_set_material_color(handle: f64, r: f64, g: f64, b: f64, a: f64) {
-    engine().scene.set_material_color(handle, r as f32, g as f32, b as f32, a as f32);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_scene_set_material_pbr(handle: f64, roughness: f64, metalness: f64) {
-    engine().scene.set_material_pbr(handle, roughness as f32, metalness as f32);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_scene_set_material_texture(handle: f64, texture_idx: f64) {
-    engine().scene.set_material_texture(handle, texture_idx as u32);
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_scene_node_count() -> f64 {
-    engine().scene.node_count() as f64
-}
-
-/// Debug: get vertex count for a scene node (0 if not found or empty)
-#[no_mangle]
-pub extern "C" fn bloom_scene_node_vertex_count(handle: f64) -> f64 {
-    match engine().scene.nodes.get(handle) {
-        Some(node) => node.vertices.len() as f64,
-        None => -1.0,
-    }
-}
-
-/// Debug: get index count for a scene node
-#[no_mangle]
-pub extern "C" fn bloom_scene_node_index_count(handle: f64) -> f64 {
-    match engine().scene.nodes.get(handle) {
-        Some(node) => node.indices.len() as f64,
-        None => -1.0,
-    }
-}
-
-
-// Q8: Set a water material on a scene node (translucent tint, low roughness).
-#[no_mangle]
-pub extern "C" fn bloom_scene_set_material_water(handle: f64, wave_amp: f64, wave_speed: f64, r: f64, g: f64, b: f64, a: f64) {
-    engine().scene.set_material_water(handle, wave_amp as f32, wave_speed as f32, r as f32, g as f32, b as f32, a as f32);
-}
-
-// Q9: Generate a ribbon mesh along a Catmull-Rom spline.
-#[cfg(feature = "models3d")]
-#[no_mangle]
-pub extern "C" fn bloom_gen_mesh_spline_ribbon(points_ptr: *const u8, point_count: f64, widths_ptr: *const u8, width_count: f64) -> f64 {
-    let n = point_count as usize;
-    let wn = width_count as usize;
-    let points = unsafe { std::slice::from_raw_parts(points_ptr as *const f32, n * 3) };
-    let widths = unsafe { std::slice::from_raw_parts(widths_ptr as *const f32, wn) };
-    engine().models.gen_mesh_spline_ribbon(points, widths)
-}
-
-// Q1: Render texture FFI — create GPU textures for render-to-texture.
-#[no_mangle]
-pub extern "C" fn bloom_load_render_texture(width: f64, height: f64) -> f64 {
-    let w = width as u32;
-    let h = height as u32;
-    let eng = engine();
-    let rt_handle = eng.textures.load_render_texture(w, h);
-
-    // Create the GPU texture via the renderer's public method.
-    let (bind_group_idx, _tex_vec_idx) = eng.renderer.create_render_texture(w, h);
-
-    // Register as a texture handle so drawTexture can sample it.
-    let tex_handle = eng.textures.textures.alloc(bloom_shared::textures::TextureData {
-        bind_group_idx, width: w, height: h,
-    });
-    eng.textures.set_render_texture_handle(rt_handle, tex_handle);
-
-    rt_handle
-}
-#[no_mangle]
-pub extern "C" fn bloom_unload_render_texture(handle: f64) {
-    engine().textures.unload_render_texture(handle);
-}
-#[no_mangle]
-pub extern "C" fn bloom_begin_texture_mode(handle: f64) {
-    let eng = engine();
-    let (w, h, bg_idx) = match eng.textures.render_textures.get(handle) {
-        Some(rt) => {
-            let tex_handle = rt.texture_handle;
-            match eng.textures.textures.get(tex_handle) {
-                Some(td) => (rt.width, rt.height, td.bind_group_idx as usize),
-                None => return,
-            }
-        }
-        None => return,
-    };
-    if let Some(texture) = eng.renderer.get_texture_ref(bg_idx) {
-        // We need to call begin_texture_mode with a reference to the texture,
-        // but get_texture_ref borrows renderer immutably. Clone the texture view
-        // data we need first, then call the mutable method.
-        let color_view = texture.create_view(&wgpu::TextureViewDescriptor::default());
-        // Create depth texture for this RT.
-        let depth_tex = eng.renderer.device.create_texture(&wgpu::TextureDescriptor {
-            label: Some("rt_depth"), size: wgpu::Extent3d { width: w, height: h, depth_or_array_layers: 1 },
-            mip_level_count: 1, sample_count: 1, dimension: wgpu::TextureDimension::D2,
-            format: wgpu::TextureFormat::Depth32Float, usage: wgpu::TextureUsages::RENDER_ATTACHMENT,
-            view_formats: &[],
-        });
-        let depth_view = depth_tex.create_view(&wgpu::TextureViewDescriptor::default());
-        eng.renderer.rt_color_view = Some(color_view);
-        eng.renderer.rt_depth_view = Some(depth_view);
-        eng.renderer.rt_depth_texture = Some(depth_tex);
-        eng.renderer.rt_width = w;
-        eng.renderer.rt_height = h;
-    }
-}
-#[no_mangle]
-pub extern "C" fn bloom_end_texture_mode() {
-    engine().renderer.end_texture_mode();
-}
-#[no_mangle]
-pub extern "C" fn bloom_get_render_texture_texture(handle: f64) -> f64 {
-    engine().textures.get_render_texture_texture(handle)
-}
 // ============================================================
 // Scene graph QoL — Q4/Q5/Q6/Q7
 // ============================================================
-
-/// Q4: Read back the 4x4 transform matrix of a scene node.
-/// Returns the column-major float at the specified index (0-15).
-#[no_mangle]
-pub extern "C" fn bloom_scene_get_transform(handle: f64, index: f64) -> f64 {
-    let mat = engine().scene.get_transform(handle);
-    let i = index as usize;
-    let col = i / 4;
-    let row = i % 4;
-    if col < 4 && row < 4 { mat[col][row] as f64 } else { 0.0 }
-}
-
-/// Q5: Get the cached AABB min of a scene node.
-#[no_mangle]
-pub extern "C" fn bloom_scene_get_bounds_min_x(handle: f64) -> f64 { engine().scene.get_bounds(handle).0[0] as f64 }
-#[no_mangle]
-pub extern "C" fn bloom_scene_get_bounds_min_y(handle: f64) -> f64 { engine().scene.get_bounds(handle).0[1] as f64 }
-#[no_mangle]
-pub extern "C" fn bloom_scene_get_bounds_min_z(handle: f64) -> f64 { engine().scene.get_bounds(handle).0[2] as f64 }
-#[no_mangle]
-pub extern "C" fn bloom_scene_get_bounds_max_x(handle: f64) -> f64 { engine().scene.get_bounds(handle).1[0] as f64 }
-#[no_mangle]
-pub extern "C" fn bloom_scene_get_bounds_max_y(handle: f64) -> f64 { engine().scene.get_bounds(handle).1[1] as f64 }
-#[no_mangle]
-pub extern "C" fn bloom_scene_get_bounds_max_z(handle: f64) -> f64 { engine().scene.get_bounds(handle).1[2] as f64 }
-
-/// Q7: Set arbitrary user data on a scene node.
-#[no_mangle]
-pub extern "C" fn bloom_scene_set_user_data(handle: f64, data: f64) {
-    engine().scene.set_user_data(handle, data as i64);
-}
-#[no_mangle]
-pub extern "C" fn bloom_scene_get_user_data(handle: f64) -> f64 {
-    engine().scene.get_user_data(handle) as f64
-}
 
 // ============================================================
 // Geometry generation
 // ============================================================
 
-/// Extrude a 2D polygon along Y axis and store as a scene node's geometry.
-/// polygon_ptr: pointer to flat f64 array [x0, z0, x1, z1, ...]
-/// polygon_count: number of points (NOT number of floats)
-/// depth: extrusion height
-#[no_mangle]
-pub extern "C" fn bloom_scene_extrude_polygon(
-    handle: f64,
-    polygon_ptr: *const f64,
-    polygon_count: f64,
-    depth: f64,
-) {
-    if polygon_ptr.is_null() { return; }
-    let n = polygon_count as usize;
-    let polygon = unsafe { std::slice::from_raw_parts(polygon_ptr, n * 2) };
-
-    let geo = bloom_shared::geometry::extrude_polygon(polygon, &[], depth);
-    engine().scene.update_geometry(handle, geo.vertices, geo.indices);
-}
-
-/// Subtract an axis-aligned box from a scene node's geometry.
-/// Removes triangles fully inside the box (simplified CSG).
-#[no_mangle]
-pub extern "C" fn bloom_scene_subtract_box(
-    handle: f64,
-    min_x: f64, min_y: f64, min_z: f64,
-    max_x: f64, max_y: f64, max_z: f64,
-) {
-    let eng = engine();
-    if let Some(node) = eng.scene.nodes.get(handle) {
-        let current = bloom_shared::geometry::GeometryData {
-            vertices: node.vertices.clone(),
-            indices: node.indices.clone(),
-        };
-        let result = bloom_shared::geometry::subtract_box(
-            &current,
-            [min_x as f32, min_y as f32, min_z as f32],
-            [max_x as f32, max_y as f32, max_z as f32],
-        );
-        eng.scene.update_geometry(handle, result.vertices, result.indices);
-    }
-}
-
 // ============================================================
 // Shadow mapping
 // ============================================================
 
-#[no_mangle]
-pub extern "C" fn bloom_enable_shadows() {
-    engine().renderer.shadow_map.enable();
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_disable_shadows() {
-    engine().renderer.shadow_map.disable();
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_dump_shadow_map(path_ptr: *const u8) {
-    let path = str_from_header(path_ptr).to_string();
-    engine().renderer.dump_shadow_map(&path);
-}
-
 // ============================================================
 // Post-processing
 // ============================================================
-
-#[no_mangle]
-pub extern "C" fn bloom_enable_postfx() {
-    let eng = engine();
-    let w = eng.renderer.width();
-    let h = eng.renderer.height();
-    let fmt = eng.renderer.surface_format();
-    eng.postfx = Some(bloom_shared::postfx::PostFxPipeline::new(
-        &eng.renderer.device, w, h, fmt,
-    ));
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_disable_postfx() {
-    engine().postfx = None;
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_postfx_set_selected(handle: f64) {
-    if let Some(pfx) = &mut engine().postfx {
-        if handle == 0.0 {
-            pfx.set_selected(Vec::new());
-        } else {
-            pfx.set_selected(vec![handle]);
-        }
-    }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_postfx_set_hovered(handle: f64) {
-    if let Some(pfx) = &mut engine().postfx {
-        pfx.set_hovered(handle);
-    }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_postfx_set_outline_color(r: f64, g: f64, b: f64, a: f64) {
-    if let Some(pfx) = &mut engine().postfx {
-        pfx.outline_params.color_selected = [r as f32, g as f32, b as f32, a as f32];
-    }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_postfx_set_outline_thickness(thickness: f64) {
-    if let Some(pfx) = &mut engine().postfx {
-        pfx.outline_params.thickness[0] = thickness as f32;
-    }
-}
 
 // ============================================================
 // 3D→2D Projection (for UI overlays positioned in 3D space)
@@ -2883,195 +951,10 @@ pub extern "C" fn bloom_postfx_set_outline_thickness(thickness: f64) {
 
 /// Project a world-space 3D point to screen coordinates.
 /// Returns screen X. Call bloom_project_y for Y. Returns -9999 if behind camera.
-static mut LAST_PROJECT: (f64, f64) = (0.0, 0.0);
-
-#[no_mangle]
-pub extern "C" fn bloom_project_to_screen(wx: f64, wy: f64, wz: f64) -> f64 {
-    let eng = engine();
-    let vp = eng.renderer.vp_matrix();
-    let w = eng.renderer.width() as f32;
-    let h = eng.renderer.height() as f32;
-
-    // Multiply by VP matrix
-    let x = wx as f32;
-    let y = wy as f32;
-    let z = wz as f32;
-    let clip_x = vp[0][0]*x + vp[1][0]*y + vp[2][0]*z + vp[3][0];
-    let clip_y = vp[0][1]*x + vp[1][1]*y + vp[2][1]*z + vp[3][1];
-    let clip_w = vp[0][3]*x + vp[1][3]*y + vp[2][3]*z + vp[3][3];
-
-    if clip_w <= 0.0 {
-        unsafe { LAST_PROJECT = (-9999.0, -9999.0); }
-        return -9999.0;
-    }
-
-    // NDC to screen
-    let ndc_x = clip_x / clip_w;
-    let ndc_y = clip_y / clip_w;
-    let screen_x = ((ndc_x + 1.0) * 0.5 * w) as f64;
-    let screen_y = ((1.0 - ndc_y) * 0.5 * h) as f64;
-
-    unsafe { LAST_PROJECT = (screen_x, screen_y); }
-    screen_x
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_project_screen_y() -> f64 {
-    unsafe { LAST_PROJECT.1 }
-}
-
-#[cfg(feature = "models3d")]
-/// Attach a loaded GLTF model's mesh geometry to a scene node.
-/// Copies the vertex/index data from the model into the scene node.
-#[no_mangle]
-pub extern "C" fn bloom_scene_attach_model(node_handle: f64, model_handle: f64, mesh_index: f64) {
-    let eng = engine();
-    let mi = mesh_index as usize;
-
-    // Get model mesh data
-    let model_data = match eng.models.models.get(model_handle) {
-        Some(md) => md,
-        None => return,
-    };
-    if mi >= model_data.meshes.len() { return; }
-    let mesh = &model_data.meshes[mi];
-
-    // Copy vertices and indices to scene node
-    let vertices = mesh.vertices.clone();
-    let indices = mesh.indices.clone();
-    let base_color_tex = mesh.texture_idx;
-    let normal_tex = mesh.normal_texture_idx;
-    let mr_tex = mesh.metallic_roughness_texture_idx;
-    let emissive_tex = mesh.emissive_texture_idx;
-    let emissive_factor = mesh.emissive_factor;
-    eng.scene.update_geometry(node_handle, vertices, indices);
-
-    // Pipe PBR textures through to the scene node material so the
-    // renderer's scene pipeline can sample them.
-    if let Some(tex_idx) = base_color_tex {
-        eng.scene.set_material_texture(node_handle, tex_idx);
-    }
-    if let Some(tex_idx) = normal_tex {
-        eng.scene.set_material_normal_texture(node_handle, tex_idx);
-    }
-    if let Some(tex_idx) = mr_tex {
-        eng.scene.set_material_metallic_roughness_texture(node_handle, tex_idx);
-    }
-    if let Some(tex_idx) = emissive_tex {
-        eng.scene.set_material_emissive_texture(node_handle, tex_idx);
-    }
-    eng.scene.set_material_emissive_factor(
-        node_handle,
-        emissive_factor[0],
-        emissive_factor[1],
-        emissive_factor[2],
-    );
-}
 
 // ============================================================
 // Thread-safe staging (for async asset loading via Perry threads)
 // ============================================================
-
-#[no_mangle]
-pub extern "C" fn bloom_stage_texture(path_ptr: *const u8) -> f64 {
-    let path = str_from_header(path_ptr);
-    match std::fs::read(path) {
-        Ok(data) => bloom_shared::staging::decode_and_stage_texture(&data),
-        Err(_) => 0.0,
-    }
-}
-
-#[cfg(feature = "models3d")]
-#[no_mangle]
-pub extern "C" fn bloom_stage_model(path_ptr: *const u8) -> f64 {
-    let path = str_from_header(path_ptr);
-    let data = match std::fs::read(path) {
-        Ok(d) => d,
-        Err(_) => return 0.0,
-    };
-    match bloom_shared::models::load_gltf_staged(&data) {
-        Some(staged) => bloom_shared::staging::stage_model(staged),
-        None => 0.0,
-    }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_stage_sound(path_ptr: *const u8) -> f64 {
-    let path = str_from_header(path_ptr);
-    let data = match std::fs::read(path) {
-        Ok(d) => d,
-        Err(_) => return 0.0,
-    };
-    let sound_data = if path.ends_with(".ogg") || path.ends_with(".OGG") {
-        parse_ogg(&data)
-    } else if path.ends_with(".mp3") || path.ends_with(".MP3") {
-        parse_mp3(&data)
-    } else {
-        parse_wav(&data)
-    };
-    match sound_data {
-        Some(sd) => bloom_shared::staging::stage_sound(sd),
-        None => 0.0,
-    }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_commit_texture(staging_handle: f64) -> f64 {
-    let staged = match bloom_shared::staging::take_texture(staging_handle) {
-        Some(s) => s,
-        None => return 0.0,
-    };
-    let eng = engine();
-    let bind_group_idx = eng.renderer.register_texture(staged.width, staged.height, &staged.data);
-    eng.textures.textures.alloc(bloom_shared::textures::TextureData {
-        bind_group_idx, width: staged.width, height: staged.height,
-    })
-}
-
-#[cfg(feature = "models3d")]
-#[no_mangle]
-pub extern "C" fn bloom_commit_model(staging_handle: f64) -> f64 {
-    let staged = match bloom_shared::staging::take_model(staging_handle) {
-        Some(s) => s,
-        None => return 0.0,
-    };
-    let eng = engine();
-    // Register staged textures with GPU and build index map.
-    // Staged texture_idx values are 1-based into staged.textures.
-    // We map them to actual renderer bind_group_idx values.
-    let mut tex_map: Vec<u32> = Vec::with_capacity(staged.textures.len());
-    for tex in &staged.textures {
-        tex_map.push(eng.renderer.register_texture(tex.width, tex.height, &tex.data));
-    }
-    let mut model = staged.model;
-    for mesh in &mut model.meshes {
-        if let Some(ref mut idx) = mesh.texture_idx {
-            let staged_idx = *idx as usize;
-            if staged_idx > 0 && staged_idx <= tex_map.len() {
-                *idx = tex_map[staged_idx - 1];
-            } else {
-                mesh.texture_idx = None;
-            }
-        }
-    }
-    eng.models.models.alloc(model)
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_commit_sound(staging_handle: f64) -> f64 {
-    match bloom_shared::staging::take_sound(staging_handle) {
-        Some(sd) => engine().audio.load_sound(sd),
-        None => 0.0,
-    }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_commit_music(staging_handle: f64) -> f64 {
-    match bloom_shared::staging::take_sound(staging_handle) {
-        Some(sd) => engine().audio.load_music(sd),
-        None => 0.0,
-    }
-}
 
 // ============================================================
 // Simple blocking executor for wgpu async calls
@@ -3293,97 +1176,8 @@ fn adler32(data: &[u8]) -> u32 {
 // Scene picking (raycasting)
 // ============================================================
 
-static mut LAST_PICK: Option<bloom_shared::picking::PickResult> = None;
-
-#[no_mangle]
-pub extern "C" fn bloom_scene_pick(screen_x: f64, screen_y: f64) -> f64 {
-    let eng = engine();
-    let inv_vp = eng.renderer.inverse_vp_matrix();
-    let cam_pos = eng.renderer.camera_pos();
-    let w = eng.renderer.width() as f32;
-    let h = eng.renderer.height() as f32;
-
-    let (origin, direction) = bloom_shared::picking::screen_to_ray(
-        screen_x as f32, screen_y as f32,
-        w, h, &inv_vp, &cam_pos,
-    );
-
-    let result = bloom_shared::picking::raycast_scene(&eng.scene, &origin, &direction);
-    let hit = result.hit;
-    unsafe { LAST_PICK = Some(result); }
-    if hit { 1.0 } else { 0.0 }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_pick_hit_handle() -> f64 {
-    unsafe { LAST_PICK.as_ref().map(|r| r.handle).unwrap_or(0.0) }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_pick_hit_distance() -> f64 {
-    unsafe { LAST_PICK.as_ref().map(|r| r.distance as f64).unwrap_or(0.0) }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_pick_hit_x() -> f64 {
-    unsafe { LAST_PICK.as_ref().map(|r| r.point[0] as f64).unwrap_or(0.0) }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_pick_hit_y() -> f64 {
-    unsafe { LAST_PICK.as_ref().map(|r| r.point[1] as f64).unwrap_or(0.0) }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_pick_hit_z() -> f64 {
-    unsafe { LAST_PICK.as_ref().map(|r| r.point[2] as f64).unwrap_or(0.0) }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_pick_hit_normal_x() -> f64 {
-    unsafe { LAST_PICK.as_ref().map(|r| r.normal[0] as f64).unwrap_or(0.0) }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_pick_hit_normal_y() -> f64 {
-    unsafe { LAST_PICK.as_ref().map(|r| r.normal[1] as f64).unwrap_or(0.0) }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_pick_hit_normal_z() -> f64 {
-    unsafe { LAST_PICK.as_ref().map(|r| r.normal[2] as f64).unwrap_or(0.0) }
-}
 
 // Q6: Multi-hit picking — returns all hits sorted by distance.
-static mut LAST_PICK_ALL: Vec<bloom_shared::picking::PickResult> = Vec::new();
-
-#[no_mangle]
-pub extern "C" fn bloom_scene_pick_all(screen_x: f64, screen_y: f64, max_results: f64) -> f64 {
-    let eng = engine();
-    let inv_vp = eng.renderer.inverse_vp_matrix();
-    let cam_pos = eng.renderer.camera_pos();
-    let w = eng.renderer.width() as f32;
-    let h = eng.renderer.height() as f32;
-    let (origin, direction) = bloom_shared::picking::screen_to_ray(
-        screen_x as f32, screen_y as f32, w, h, &inv_vp, &cam_pos,
-    );
-    let results = bloom_shared::picking::raycast_scene_all(&eng.scene, &origin, &direction, max_results as usize);
-    let count = results.len();
-    unsafe { LAST_PICK_ALL = results; }
-    count as f64
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_pick_all_handle(index: f64) -> f64 {
-    let i = index as usize;
-    unsafe { LAST_PICK_ALL.get(i).map(|r| r.handle).unwrap_or(0.0) }
-}
-
-#[no_mangle]
-pub extern "C" fn bloom_pick_all_distance(index: f64) -> f64 {
-    let i = index as usize;
-    unsafe { LAST_PICK_ALL.get(i).map(|r| r.distance as f64).unwrap_or(0.0) }
-}
 
 // ============================================================
 // Physics (Jolt 5.x) — FFI surface generated from shared macro
@@ -3398,170 +1192,3 @@ fn bloom_jolt_ffi_physics() -> &'static mut bloom_shared::physics_jolt::JoltPhys
 #[cfg(feature = "jolt")]
 bloom_shared::define_physics_ffi!();
 
-// ─── EN-014 feature-off stubs ───────────────────────────────────────
-// The Perry TS glue for an imported engine module references every FFI
-// symbol that module declares, whether or not the game calls it. When a
-// feature is compiled out, keep the symbols as warn-once no-ops so the
-// link succeeds and a stray call is diagnosable instead of UB.
-#[cfg(not(feature = "image-extras"))]
-#[no_mangle]
-pub extern "C" fn bloom_set_env_clear_from_hdr(_path_ptr: *const u8) {
-    feature_off_warn_once("bloom_set_env_clear_from_hdr", "image-extras");
-}
-#[cfg(not(feature = "models3d"))]
-#[no_mangle]
-pub extern "C" fn bloom_load_model(_path_ptr: *const u8) -> f64 {
-    feature_off_warn_once("bloom_load_model", "models3d");
-    0.0
-}
-#[cfg(not(feature = "models3d"))]
-#[no_mangle]
-pub extern "C" fn bloom_draw_model(_handle: f64, _x: f64, _y: f64, _z: f64, _scale: f64, _r: f64, _g: f64, _b: f64, _a: f64) {
-    feature_off_warn_once("bloom_draw_model", "models3d");
-}
-#[cfg(not(feature = "models3d"))]
-#[no_mangle]
-pub extern "C" fn bloom_draw_model_rotated(
-    handle: f64, _x: f64, _y: f64, _z: f64,
-    _scale: f64, _rot_y: f64,
-    _color_packed_argb: f64,
-) {
-    feature_off_warn_once("bloom_draw_model_rotated", "models3d");
-}
-#[cfg(not(feature = "models3d"))]
-#[no_mangle]
-pub extern "C" fn bloom_unload_model(_handle: f64) {
-    feature_off_warn_once("bloom_unload_model", "models3d");
-}
-#[cfg(not(feature = "models3d"))]
-#[no_mangle]
-pub extern "C" fn bloom_gen_mesh_cube(_w: f64, _h: f64, _d: f64) -> f64 {
-    feature_off_warn_once("bloom_gen_mesh_cube", "models3d");
-    0.0
-}
-#[cfg(not(feature = "models3d"))]
-#[no_mangle]
-pub extern "C" fn bloom_gen_mesh_heightmap(_image_handle: f64, _size_x: f64, _size_y: f64, _size_z: f64) -> f64 {
-    feature_off_warn_once("bloom_gen_mesh_heightmap", "models3d");
-    0.0
-}
-#[cfg(not(feature = "models3d"))]
-#[no_mangle]
-pub extern "C" fn bloom_draw_material(
-    material: f64,
-    _mesh_handle: f64,
-    _mesh_idx: f64,
-    _x: f64, _y: f64, _z: f64, _scale: f64,
-    _r: f64, _g: f64, _b: f64, _a: f64,
-) {
-    feature_off_warn_once("bloom_draw_material", "models3d");
-}
-#[cfg(not(feature = "models3d"))]
-#[no_mangle]
-pub extern "C" fn bloom_create_mesh(_vertex_ptr: *const f64, _vertex_count: f64, _index_ptr: *const f64, _index_count: f64) -> f64 {
-    feature_off_warn_once("bloom_create_mesh", "models3d");
-    0.0
-}
-#[cfg(not(feature = "models3d"))]
-#[no_mangle]
-pub extern "C" fn bloom_load_model_animation(_path_ptr: *const u8) -> f64 {
-    feature_off_warn_once("bloom_load_model_animation", "models3d");
-    0.0
-}
-#[cfg(not(feature = "models3d"))]
-#[no_mangle]
-pub extern "C" fn bloom_update_model_animation(_handle: f64, _anim_index: f64, _time: f64, _scale: f64, _px: f64, _py: f64, _pz: f64, _rot_y: f64) {
-    feature_off_warn_once("bloom_update_model_animation", "models3d");
-}
-#[cfg(not(feature = "models3d"))]
-#[no_mangle]
-pub extern "C" fn bloom_get_model_mesh_count(_handle: f64) -> f64 {
-    feature_off_warn_once("bloom_get_model_mesh_count", "models3d");
-    0.0
-}
-#[cfg(not(feature = "models3d"))]
-#[no_mangle]
-pub extern "C" fn bloom_get_model_material_count(_handle: f64) -> f64 {
-    feature_off_warn_once("bloom_get_model_material_count", "models3d");
-    0.0
-}
-#[cfg(not(feature = "models3d"))]
-#[no_mangle]
-pub extern "C" fn bloom_get_model_bounds_min_x(_model_handle: f64) -> f64 {
-    feature_off_warn_once("bloom_get_model_bounds_min_x", "models3d");
-    0.0
-}
-#[cfg(not(feature = "models3d"))]
-#[no_mangle]
-pub extern "C" fn bloom_get_model_bounds_min_y(_model_handle: f64) -> f64 {
-    feature_off_warn_once("bloom_get_model_bounds_min_y", "models3d");
-    0.0
-}
-#[cfg(not(feature = "models3d"))]
-#[no_mangle]
-pub extern "C" fn bloom_get_model_bounds_min_z(_model_handle: f64) -> f64 {
-    feature_off_warn_once("bloom_get_model_bounds_min_z", "models3d");
-    0.0
-}
-#[cfg(not(feature = "models3d"))]
-#[no_mangle]
-pub extern "C" fn bloom_get_model_bounds_max_x(_model_handle: f64) -> f64 {
-    feature_off_warn_once("bloom_get_model_bounds_max_x", "models3d");
-    0.0
-}
-#[cfg(not(feature = "models3d"))]
-#[no_mangle]
-pub extern "C" fn bloom_get_model_bounds_max_y(_model_handle: f64) -> f64 {
-    feature_off_warn_once("bloom_get_model_bounds_max_y", "models3d");
-    0.0
-}
-#[cfg(not(feature = "models3d"))]
-#[no_mangle]
-pub extern "C" fn bloom_get_model_bounds_max_z(_model_handle: f64) -> f64 {
-    feature_off_warn_once("bloom_get_model_bounds_max_z", "models3d");
-    0.0
-}
-#[cfg(not(feature = "models3d"))]
-#[no_mangle]
-pub extern "C" fn bloom_gen_mesh_spline_ribbon(_points_ptr: *const u8, _point_count: f64, _widths_ptr: *const u8, _width_count: f64) -> f64 {
-    feature_off_warn_once("bloom_gen_mesh_spline_ribbon", "models3d");
-    0.0
-}
-#[cfg(not(feature = "models3d"))]
-#[no_mangle]
-pub extern "C" fn bloom_scene_attach_model(_node_handle: f64, _model_handle: f64, _mesh_index: f64) {
-    feature_off_warn_once("bloom_scene_attach_model", "models3d");
-}
-#[cfg(not(feature = "models3d"))]
-#[no_mangle]
-pub extern "C" fn bloom_stage_model(_path_ptr: *const u8) -> f64 {
-    feature_off_warn_once("bloom_stage_model", "models3d");
-    0.0
-}
-#[cfg(not(feature = "models3d"))]
-#[no_mangle]
-pub extern "C" fn bloom_commit_model(_staging_handle: f64) -> f64 {
-    feature_off_warn_once("bloom_commit_model", "models3d");
-    0.0
-}
-
-#[cfg(any(not(feature = "models3d"), not(feature = "image-extras")))]
-fn feature_off_warn_once(name: &str, feature: &str) {
-    use std::sync::Mutex;
-    static WARNED: Mutex<Vec<&'static str>> = Mutex::new(Vec::new());
-    let mut warned = WARNED.lock().unwrap();
-    if !warned.iter().any(|n| *n == name) {
-        // names come from string literals above — leak-free 'static via Box::leak is
-        // overkill; just store a leaked copy once per distinct symbol.
-        warned.push(Box::leak(name.to_string().into_boxed_str()));
-        eprintln!("bloom: {name}() ignored — engine built without the `{feature}` feature");
-    }
-}
-#[cfg(not(feature = "models3d"))]
-#[no_mangle]
-pub extern "C" fn bloom_submit_material_draw_instanced(
-    _material: f64, _mesh_handle: f64, _mesh_idx: f64,
-    _instance_buffer: f64, _instance_count: f64,
-) {
-    feature_off_warn_once("bloom_submit_material_draw_instanced", "models3d");
-}
