+++ native/shared/src/ffi.rs
@@ -1,0 +1,109 @@
+//! FFI boundary infrastructure shared by every platform crate.
+//!
+//! All `bloom_*` extern "C" functions generated by [`define_core_ffi!`]
+//! (and, over time, the hand-written platform-specific ones) route their
+//! bodies through [`guard`], which:
+//!
+//!   1. Prevents a Rust panic from unwinding across the `extern "C"`
+//!      boundary into Perry-compiled code. Release builds use
+//!      `panic = "abort"` where this is moot, but dev/test profiles unwind,
+//!      and rustc only aborts at the boundary since 1.81 — logging the
+//!      function name first is strictly more debuggable than a bare abort.
+//!   2. Converts a panic into a logged, once-per-symbol error plus a safe
+//!      default return value, so a single bad call (stale handle, NaN
+//!      argument, asset race) degrades that call instead of killing the
+//!      game process.
+//!
+//! Logging goes through [`log_error`], which writes to logcat on Android
+//! (stderr is discarded there) and stderr everywhere else.
+
+/// Default return value for an FFI function whose body panicked.
+pub trait FfiDefault {
+    fn ffi_default() -> Self;
+}
+
+impl FfiDefault for () {
+    fn ffi_default() -> Self {}
+}
+
+impl FfiDefault for f64 {
+    fn ffi_default() -> Self {
+        0.0
+    }
+}
+
+impl FfiDefault for *const u8 {
+    fn ffi_default() -> Self {
+        std::ptr::null()
+    }
+}
+
+/// Run an FFI function body, catching panics at the boundary.
+///
+/// On panic: logs `name` + the panic payload (once per symbol, so a
+/// panicking per-frame call doesn't flood the log) and returns
+/// `T::ffi_default()`.
+pub fn guard<T: FfiDefault>(name: &'static str, f: impl FnOnce() -> T) -> T {
+    match std::panic::catch_unwind(std::panic::AssertUnwindSafe(f)) {
+        Ok(v) => v,
+        Err(payload) => {
+            warn_once(name, || {
+                let msg = if let Some(s) = payload.downcast_ref::<&str>() {
+                    (*s).to_string()
+                } else if let Some(s) = payload.downcast_ref::<String>() {
+                    s.clone()
+                } else {
+                    "non-string panic payload".to_string()
+                };
+                format!("bloom: panic in {name}(): {msg} — returning default; further panics in this function are suppressed")
+            });
+            T::ffi_default()
+        }
+    }
+}
+
+/// Log `make_msg()` the first time `key` is seen, silently no-op after.
+fn warn_once(key: &'static str, make_msg: impl FnOnce() -> String) {
+    use std::sync::Mutex;
+    static SEEN: Mutex<Vec<&'static str>> = Mutex::new(Vec::new());
+    let mut seen = match SEEN.lock() {
+        Ok(g) => g,
+        // Poisoned only if a previous caller panicked while holding the
+        // lock; logging is best-effort, keep going with the inner data.
+        Err(p) => p.into_inner(),
+    };
+    if !seen.contains(&key) {
+        seen.push(key);
+        log_error(&make_msg());
+    }
+}
+
+/// One-time "this function is compiled out" notice for feature-gated FFI
+/// stubs (models3d, image-extras). A silent stub here cost weeks of
+/// confusion historically (Windows shipped silent no-ops for the whole
+/// scene-graph surface) — never strip the warning.
+pub fn feature_off_warn_once(name: &'static str, feature: &'static str) {
+    warn_once(name, || {
+        format!("bloom: {name}() ignored — engine built without the `{feature}` cargo feature")
+    });
+}
+
+/// Platform-appropriate error log: logcat on Android, stderr elsewhere.
+pub fn log_error(msg: &str) {
+    #[cfg(target_os = "android")]
+    {
+        // liblog is linked by every Android app process.
+        extern "C" {
+            fn __android_log_print(prio: i32, tag: *const u8, fmt: *const u8, ...) -> i32;
+        }
+        if let Ok(c) = std::ffi::CString::new(msg) {
+            // 6 = ANDROID_LOG_ERROR
+            unsafe {
+                __android_log_print(6, b"BloomEngine\0".as_ptr(), b"%s\0".as_ptr(), c.as_ptr());
+            }
+        }
+        return;
+    }
+    #[cfg(not(target_os = "android"))]
+    eprintln!("{msg}");
+}
